chore: refactor sovd manager into modules.
This is the first commit in the diag_kernel refactoring,
to split the module up.
It extracts test_utils module which is shared among the different
modules to create ecumanager instances and mock mdd datatypes.
Service lookups and comparams are also refactored into a new module.
