+++ cda-core/src/diag_kernel/ecumanager.rs
@@ -15,9 +15,9 @@ use std::{sync::Arc, time::Duration};
 
 use cda_database::datatypes;
 use cda_interfaces::{
-    DiagComm, DiagCommAction, DiagCommType, DiagServiceError, DynamicPlugin, EcuManagerType,
-    EcuState, EcuVariant, HashMap, HashMapExtensions, HashSet, HashSetExtensions, Protocol,
-    STRINGS, SecurityAccess, ServicePayload, StringId,
+    DiagComm, DiagCommType, DiagServiceError, DynamicPlugin, EcuManagerType, EcuState, EcuVariant,
+    HashMap, HashMapExtensions, HashSet, HashSetExtensions, Protocol, SecurityAccess,
+    ServicePayload,
     datatypes::{
         AddressingMode, CLEAR_FAULT_MEM_POS_RESPONSE_SID, ComParamConfig, ComParamPrecedence,
         ComParams, ComplexComParamValue, ComponentConfigurationsInfo, ComponentDataInfo,
@@ -35,6 +35,7 @@ use cda_interfaces::{
 use cda_plugin_security::SecurityPlugin;
 use tokio::sync::RwLock;
 
+use super::service_lookup::DbCache;
 use crate::{
     DiagDataContainerDtc, MappedResponseData,
     diag_kernel::{
@@ -128,73 +129,57 @@ pub struct EcuManagerConfig {
 #[allow(clippy::struct_excessive_bools)]
 pub struct EcuManager<S: SecurityPlugin> {
     pub(crate) diag_database: datatypes::DiagnosticDatabase,
-    db_cache: DbCache,
-    ecu_name: String,
-    description_type: EcuManagerType,
-    database_naming_convention: DatabaseNamingConvention,
-    tester_address: u16,
-    logical_address: u16,
-    logical_gateway_address: u16,
-    logical_functional_address: u16,
+    pub(crate) db_cache: DbCache,
+    pub(crate) ecu_name: String,
+    pub(crate) description_type: EcuManagerType,
+    pub(crate) database_naming_convention: DatabaseNamingConvention,
+    pub(crate) tester_address: u16,
+    pub(crate) logical_address: u16,
+    pub(crate) logical_gateway_address: u16,
+    pub(crate) logical_functional_address: u16,
 
-    nack_number_of_retries: HashMap<u8, u32>,
-    diagnostic_ack_timeout: Duration,
-    retry_period: Duration,
-    routing_activation_timeout: Duration,
-    repeat_request_count_transmission: u32,
-    connection_timeout: Duration,
-    connection_retry_delay: Duration,
-    connection_retry_attempts: u32,
+    pub(crate) nack_number_of_retries: HashMap<u8, u32>,
+    pub(crate) diagnostic_ack_timeout: Duration,
+    pub(crate) retry_period: Duration,
+    pub(crate) routing_activation_timeout: Duration,
+    pub(crate) repeat_request_count_transmission: u32,
+    pub(crate) connection_timeout: Duration,
+    pub(crate) connection_retry_delay: Duration,
+    pub(crate) connection_retry_attempts: u32,
 
     variant_detection: variant_detection::VariantDetection,
-    variant_index: Option<usize>,
-    variant: EcuVariant,
-    fallback_to_base_variant: bool,
-    strict_parameter_validation: bool,
-    duplicating_ecu_names: Option<HashSet<String>>,
+    pub(crate) variant_index: Option<usize>,
+    pub(crate) variant: EcuVariant,
+    pub(crate) fallback_to_base_variant: bool,
+    pub(crate) strict_parameter_validation: bool,
+    pub(crate) duplicating_ecu_names: Option<HashSet<String>>,
 
-    protocol: Protocol,
+    pub(crate) protocol: Protocol,
     // functional group: protocol prefixed or postfixed
-    fg_protocol_position: DiagnosticServiceAffixPosition,
-    ecu_service_states: Arc<RwLock<HashMap<u8, String>>>,
+    pub(crate) fg_protocol_position: DiagnosticServiceAffixPosition,
+    pub(crate) ecu_service_states: Arc<RwLock<HashMap<u8, String>>>,
 
-    tester_present_retry_policy: bool,
-    tester_present_addr_mode: AddressingMode,
-    tester_present_response_expected: bool,
-    tester_present_send_type: TesterPresentSendType,
-    tester_present_message: Vec<u8>,
-    tester_present_exp_pos_resp: Vec<u8>,
-    tester_present_exp_neg_resp: Vec<u8>,
-    tester_present_time: Duration,
-    repeat_req_count_app: u32,
-    rc_21_retry_policy: RetryPolicy,
-    rc_21_completion_timeout: Duration,
-    rc_21_repeat_request_time: Duration,
-    rc_78_retry_policy: RetryPolicy,
-    rc_78_completion_timeout: Duration,
-    rc_78_timeout: Duration,
-    rc_94_retry_policy: RetryPolicy,
-    rc_94_completion_timeout: Duration,
-    rc_94_repeat_request_time: Duration,
-    timeout_default: Duration,
+    pub(crate) tester_present_retry_policy: bool,
+    pub(crate) tester_present_addr_mode: AddressingMode,
+    pub(crate) tester_present_response_expected: bool,
+    pub(crate) tester_present_send_type: TesterPresentSendType,
+    pub(crate) tester_present_message: Vec<u8>,
+    pub(crate) tester_present_exp_pos_resp: Vec<u8>,
+    pub(crate) tester_present_exp_neg_resp: Vec<u8>,
+    pub(crate) tester_present_time: Duration,
+    pub(crate) repeat_req_count_app: u32,
+    pub(crate) rc_21_retry_policy: RetryPolicy,
+    pub(crate) rc_21_completion_timeout: Duration,
+    pub(crate) rc_21_repeat_request_time: Duration,
+    pub(crate) rc_78_retry_policy: RetryPolicy,
+    pub(crate) rc_78_completion_timeout: Duration,
+    pub(crate) rc_78_timeout: Duration,
+    pub(crate) rc_94_retry_policy: RetryPolicy,
+    pub(crate) rc_94_completion_timeout: Duration,
+    pub(crate) rc_94_repeat_request_time: Duration,
+    pub(crate) timeout_default: Duration,
 
     security_plugin_phantom: std::marker::PhantomData<S>,
-}
-
-#[derive(Default)]
-struct DbCache {
-    pub(crate) diag_services: RwLock<HashMap<StringId, Option<CacheLocation>>>,
-}
-
-impl DbCache {
-    pub(crate) async fn reset(&self) {
-        self.diag_services.write().await.clear();
-    }
-}
-
-enum CacheLocation {
-    Variant(usize),
-    ParentRef(usize),
 }
 
 impl<S: SecurityPlugin> cda_interfaces::EcuAddressProvider for EcuManager<S> {
@@ -1881,93 +1866,6 @@ impl<S: SecurityPlugin> cda_interfaces::EcuManager for EcuManager<S> {
     }
 }
 
-impl<S: SecurityPlugin> cda_interfaces::UdsComParamProvider for EcuManager<S> {
-    fn tester_present_retry_policy(&self) -> bool {
-        self.tester_present_retry_policy
-    }
-    fn tester_present_addr_mode(self) -> AddressingMode {
-        self.tester_present_addr_mode.clone()
-    }
-    fn tester_present_response_expected(self) -> bool {
-        self.tester_present_response_expected
-    }
-    fn tester_present_send_type(self) -> TesterPresentSendType {
-        self.tester_present_send_type.clone()
-    }
-    fn tester_present_message(self) -> Vec<u8> {
-        self.tester_present_message.clone()
-    }
-    fn tester_present_exp_pos_resp(self) -> Vec<u8> {
-        self.tester_present_exp_pos_resp.clone()
-    }
-    fn tester_present_exp_neg_resp(self) -> Vec<u8> {
-        self.tester_present_exp_neg_resp.clone()
-    }
-    fn tester_present_time(&self) -> Duration {
-        self.tester_present_time
-    }
-    fn repeat_req_count_app(&self) -> u32 {
-        self.repeat_req_count_app
-    }
-    fn rc_21_retry_policy(&self) -> RetryPolicy {
-        self.rc_21_retry_policy.clone()
-    }
-    fn rc_21_completion_timeout(&self) -> Duration {
-        self.rc_21_completion_timeout
-    }
-    fn rc_21_repeat_request_time(&self) -> Duration {
-        self.rc_21_repeat_request_time
-    }
-    fn rc_78_retry_policy(&self) -> RetryPolicy {
-        self.rc_78_retry_policy.clone()
-    }
-    fn rc_78_completion_timeout(&self) -> Duration {
-        self.rc_78_completion_timeout
-    }
-    fn rc_78_timeout(&self) -> Duration {
-        self.rc_78_timeout
-    }
-    fn rc_94_retry_policy(&self) -> RetryPolicy {
-        self.rc_94_retry_policy.clone()
-    }
-    fn rc_94_completion_timeout(&self) -> Duration {
-        self.rc_94_completion_timeout
-    }
-    fn rc_94_repeat_request_time(&self) -> Duration {
-        self.rc_94_repeat_request_time
-    }
-    fn timeout_default(&self) -> Duration {
-        self.timeout_default
-    }
-}
-
-impl<S: SecurityPlugin> cda_interfaces::DoipComParamProvider for EcuManager<S> {
-    fn nack_number_of_retries(&self) -> &HashMap<u8, u32> {
-        &self.nack_number_of_retries
-    }
-    fn diagnostic_ack_timeout(&self) -> Duration {
-        self.diagnostic_ack_timeout
-    }
-    fn retry_period(&self) -> Duration {
-        self.retry_period
-    }
-    fn routing_activation_timeout(&self) -> Duration {
-        self.routing_activation_timeout
-    }
-    fn repeat_request_count_transmission(&self) -> u32 {
-        self.repeat_request_count_transmission
-    }
-    fn connection_timeout(&self) -> Duration {
-        self.connection_timeout
-    }
-    fn connection_retry_delay(&self) -> Duration {
-        self.connection_retry_delay
-    }
-    fn connection_retry_attempts(&self) -> u32 {
-        self.connection_retry_attempts
-    }
-}
-
 impl<S: SecurityPlugin> EcuManager<S> {
     /// Trims affixes from a routine control service name to derive the base routine name.
     fn trim_routine_name(&self, name: &str) -> String {
@@ -2374,7 +2272,7 @@ impl<S: SecurityPlugin> EcuManager<S> {
         Ok(())
     }
 
-    fn variant(&self) -> Option<datatypes::Variant<'_>> {
+    pub(crate) fn variant(&self) -> Option<datatypes::Variant<'_>> {
         let idx = self.variant_index?;
         let variants = self.diag_database.ecu_data().ok()?.variants()?;
         Some(variants.get(idx).into())
@@ -2396,667 +2294,6 @@ impl<S: SecurityPlugin> EcuManager<S> {
                     self.database_naming_convention.trim_long_name_affixes(v)
                 }),
         }
-    }
-
-    /// Lookup a diagnostic service by its diag comm definition.
-    ///
-    /// When `functional_group_name` is `Some`, the service is looked up in the
-    /// named functional group's `DiagLayer` and its parent references instead of
-    /// the ECU variant. When `None`, the lookup uses the ECU variant (with
-    /// caching, because it is used for *every* UDS request).
-    ///
-    /// `subfunction_mask` is an optional bitmask applied to both the incoming
-    /// `DiagComm::subfunction_id` and the database service's subfunction value
-    /// before comparing.  When it is `None`, [`DEFAULT_SUBFUNCTION_MASK`] (`0x7F`) is
-    /// used, which masks out the suppress-positive-response bit (bit 7).
-    pub(in crate::diag_kernel) async fn lookup_diag_service(
-        &self,
-        diag_comm: &cda_interfaces::DiagComm,
-        functional_group_name: Option<&str>,
-        subfunction_mask: Option<u8>,
-    ) -> Result<datatypes::DiagService<'_>, DiagServiceError> {
-        // When a subfunction id is provided, match on base name + subfunction id
-        // rather than constructing a name suffix. The cache key encodes both so
-        // that a subfunction-id lookup and a plain name lookup for the same base
-        // name never collide.
-        if let Some(sf_id) = diag_comm.subfunction_id {
-            let base_name = diag_comm.name.to_lowercase();
-            let effective_mask =
-                subfunction_mask.unwrap_or(cda_interfaces::DEFAULT_SUBFUNCTION_MASK);
-            let mask_u32 = u32::from(effective_mask);
-
-            let prefixes = diag_comm.type_.service_prefixes();
-            let predicate = |service: &datatypes::DiagService<'_>| {
-                service.diag_comm().is_some_and(|dc| {
-                    dc.short_name()
-                        .is_some_and(|name| starts_with_ignore_ascii_case(name, &base_name))
-                }) && service
-                    .request_id()
-                    .is_some_and(|sid| prefixes.contains(&sid))
-                    && service
-                        .request_sub_function_id()
-                        .is_some_and(|(id, _)| (id & mask_u32) == (u32::from(sf_id) & mask_u32))
-            };
-
-            if let Some(fg_name) = functional_group_name {
-                return self
-                    .get_services_from_functional_group_and_parent_refs(fg_name, predicate)?
-                    .into_iter()
-                    .next()
-                    .ok_or_else(|| {
-                        DiagServiceError::NotFound(format!(
-                            "Diagnostic service '{base_name}' with subfunction {sf_id:#04X} not \
-                             found in functional group '{fg_name}'"
-                        ))
-                    });
-            }
-
-            let cache_key = format!("{base_name}:sf{sf_id}:m{effective_mask:02X}");
-            let lookup_id = STRINGS.get_or_insert(&cache_key);
-
-            if let Some(Some(location)) = self.db_cache.diag_services.read().await.get(&lookup_id) {
-                return match self.get_service_by_location(location) {
-                    Some(service) => Ok(service),
-                    None => Err(DiagServiceError::NotFound(format!(
-                        "Cached diagnostic service '{base_name}' with subfunction {sf_id:#04X} \
-                         not found at stored location"
-                    ))),
-                };
-            }
-
-            if let Some((service, location)) = self.search_with_location(&predicate) {
-                self.db_cache
-                    .diag_services
-                    .write()
-                    .await
-                    .insert(lookup_id, Some(location));
-                return Ok(service);
-            }
-
-            self.db_cache
-                .diag_services
-                .write()
-                .await
-                .insert(lookup_id, None);
-
-            return Err(DiagServiceError::NotFound(format!(
-                "Diagnostic service '{base_name}' with subfunction {sf_id:#04X} not found in \
-                 variant or parent refs"
-            )));
-        }
-
-        let lookup_name = if let Some(name) = &diag_comm.lookup_name {
-            name.to_owned()
-        } else {
-            match diag_comm.action() {
-                DiagCommAction::Read => format!("{}_Read", diag_comm.name),
-                DiagCommAction::Write => format!("{}_Write", diag_comm.name),
-                DiagCommAction::Start => format!("{}_Start", diag_comm.name),
-                DiagCommAction::RequestResults => {
-                    format!("{}_RequestResults", diag_comm.name)
-                }
-                DiagCommAction::Stop => format!("{}_Stop", diag_comm.name),
-            }
-        }
-        .to_lowercase();
-
-        let prefixes = diag_comm.type_.service_prefixes();
-        let predicate = |service: &datatypes::DiagService<'_>| {
-            service.diag_comm().is_some_and(|dc| {
-                dc.short_name()
-                    .is_some_and(|name| starts_with_ignore_ascii_case(name, &lookup_name))
-            }) && service
-                .request_id()
-                .is_some_and(|sid| prefixes.contains(&sid))
-        };
-
-        if let Some(fg_name) = functional_group_name {
-            return self
-                .get_services_from_functional_group_and_parent_refs(fg_name, predicate)?
-                .into_iter()
-                .next()
-                .ok_or_else(|| {
-                    DiagServiceError::NotFound(format!(
-                        "Diagnostic service '{lookup_name}' not found in functional group \
-                         '{fg_name}'"
-                    ))
-                });
-        }
-
-        let lookup_id = STRINGS.get_or_insert(&lookup_name);
-
-        if let Some(Some(location)) = self.db_cache.diag_services.read().await.get(&lookup_id) {
-            return match self.get_service_by_location(location) {
-                Some(service) => Ok(service),
-                None => Err(DiagServiceError::NotFound(format!(
-                    "Cached diagnostic service '{lookup_name}' not found at stored location"
-                ))),
-            };
-        }
-
-        // Search and cache the location
-        if let Some((service, location)) = self.search_with_location(&predicate) {
-            self.db_cache
-                .diag_services
-                .write()
-                .await
-                .insert(lookup_id, Some(location));
-            return Ok(service);
-        }
-
-        self.db_cache
-            .diag_services
-            .write()
-            .await
-            .insert(lookup_id, None);
-
-        Err(DiagServiceError::NotFound(format!(
-            "Diagnostic service '{lookup_name}' not found in variant, base variant, or ECU shared \
-             data"
-        )))
-    }
-
-    fn search_with_location<F>(
-        &self,
-        predicate: &F,
-    ) -> Option<(datatypes::DiagService<'_>, CacheLocation)>
-    where
-        F: Fn(&datatypes::DiagService<'_>) -> bool,
-    {
-        // Search in variant
-        if let Some((idx, service)) = self
-            .variant()
-            // This is necessary, so we are able to lookup services
-            // _before_ a variant has been found i.e. for variant detection.
-            .or_else(|| self.diag_database.base_variant().ok())
-            .and_then(|v| v.diag_layer())
-            .and_then(|dl| dl.diag_services())
-            .and_then(|services| {
-                services.iter().enumerate().find_map(|(idx, s)| {
-                    let service = datatypes::DiagService(s);
-                    predicate(&service).then_some((idx, service))
-                })
-            })
-        {
-            return Some((service, CacheLocation::Variant(idx)));
-        }
-
-        // Search in Parent Refs
-        if let Some((idx, service)) = self.get_variant_parent_ref_services().and_then(|services| {
-            services
-                .iter()
-                .enumerate()
-                .find_map(|(idx, s)| predicate(s).then_some((idx, s.clone())))
-        }) {
-            return Some((service, CacheLocation::ParentRef(idx)));
-        }
-
-        None
-    }
-
-    fn get_services_from_diag_layer_and_parent_refs<'a, F>(
-        diag_layer: &datatypes::DiagLayer<'a>,
-        parent_refs: impl Iterator<Item = impl Into<datatypes::ParentRef<'a>>>,
-        service_filter: F,
-    ) -> Vec<datatypes::DiagService<'a>>
-    where
-        F: Fn(&datatypes::DiagService) -> bool,
-    {
-        diag_layer
-            .diag_services()
-            .into_iter()
-            .flatten()
-            .map(datatypes::DiagService)
-            .chain(
-                Self::get_parent_ref_services_recursive(parent_refs)
-                    .into_iter()
-                    .flatten(),
-            )
-            .filter(service_filter)
-            .collect()
-    }
-
-    /// Retrieves single ECU jobs from a given `DiagLayer` and its parent references,
-    /// filtered by the provided predicate. Jobs from the `DiagLayer` are returned first,
-    /// followed by jobs resolved recursively from parent references.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let jobs = EcuManager::get_single_ecu_jobs_from_diag_layer_and_parent_refs(
-    ///     &diag_layer,
-    ///     parent_refs.into_iter().map(datatypes::ParentRef),
-    ///     |job| job.diag_comm().and_then(|dc| dc.short_name()) == Some("MyJob"),
-    /// );
-    /// ```
-    fn get_single_ecu_jobs_from_diag_layer_and_parent_refs<'a, F>(
-        diag_layer: &datatypes::DiagLayer<'a>,
-        parent_refs: impl Iterator<Item = impl Into<datatypes::ParentRef<'a>>>,
-        service_filter: F,
-    ) -> Vec<datatypes::SingleEcuJob<'a>>
-    where
-        F: Fn(&datatypes::SingleEcuJob) -> bool,
-    {
-        diag_layer
-            .single_ecu_jobs()
-            .into_iter()
-            .flatten()
-            .map(datatypes::SingleEcuJob)
-            .chain(
-                Self::get_parent_ref_jobs_recursive(parent_refs)
-                    .into_iter()
-                    .flatten(),
-            )
-            .filter(service_filter)
-            .collect()
-    }
-
-    /// Retrieves diagnostic services from the current variants `DiagLayer` and its parent
-    /// references, filtered by the provided predicate. Returns an empty vector if no variant
-    /// is set.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let read_services = ecu_manager.get_services_from_variant_and_parent_refs(|service| {
-    ///     service
-    ///         .request_id()
-    ///         .is_some_and(|id| id == service_ids::READ_DATA_BY_IDENTIFIER)
-    /// });
-    /// for service in &read_services {
-    ///     println!("{:?}", service.diag_comm().and_then(|dc| dc.short_name()));
-    /// }
-    /// ```
-    fn get_services_from_variant_and_parent_refs<F>(
-        &self,
-        service_filter: F,
-    ) -> Vec<datatypes::DiagService<'_>>
-    where
-        F: Fn(&datatypes::DiagService) -> bool,
-    {
-        self.variant()
-            .and_then(|v| v.diag_layer().map(|dl| (dl, v.parent_refs())))
-            .map_or(<_>::default(), |(diag_layer, parent_refs)| {
-                Self::get_services_from_diag_layer_and_parent_refs(
-                    &(diag_layer.into()),
-                    parent_refs.into_iter().flatten().map(datatypes::ParentRef),
-                    service_filter,
-                )
-            })
-    }
-
-    /// Retrieves diagnostic services from a given functional group and its parent
-    /// references, filtered by the provided predicate.
-    ///
-    /// # Errors
-    /// Will return `Err` if the database has no functional groups or the specified
-    /// group is not found.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let read_services = ecu_manager.get_services_from_functional_group_and_parent_refs(
-    ///     "FunctionalGroupName",
-    ///     |service| {
-    ///         service
-    ///             .request_id()
-    ///             .is_some_and(|id| id == service_ids::READ_DATA_BY_IDENTIFIER)
-    ///     },
-    /// )?;
-    /// for service in &read_services {
-    ///     println!("{:?}", service.diag_comm().and_then(|dc| dc.short_name()));
-    /// }
-    /// ```
-    fn get_services_from_functional_group_and_parent_refs<F>(
-        &self,
-        group_name: &str,
-        service_filter: F,
-    ) -> Result<Vec<datatypes::DiagService<'_>>, DiagServiceError>
-    where
-        F: Fn(&datatypes::DiagService) -> bool,
-    {
-        let Ok(groups) = self.diag_database.functional_groups() else {
-            return Err(DiagServiceError::InvalidDatabase(
-                "Database has no functional groups".to_owned(),
-            ));
-        };
-
-        let matching_group = groups
-            .into_iter()
-            .find(|group| {
-                group
-                    .diag_layer()
-                    .and_then(|dl| dl.short_name())
-                    .is_some_and(|name| name.eq_ignore_ascii_case(group_name))
-            })
-            .ok_or_else(|| {
-                DiagServiceError::NotFound(format!("Functional group '{group_name}' not found"))
-            })?;
-
-        Ok(matching_group
-            .diag_layer()
-            .map(|dl| (dl, matching_group.parent_refs()))
-            .map_or(<_>::default(), |(diag_layer, parent_refs)| {
-                Self::get_services_from_diag_layer_and_parent_refs(
-                    &(diag_layer.into()),
-                    parent_refs.into_iter().flatten().map(datatypes::ParentRef),
-                    service_filter,
-                )
-            }))
-    }
-
-    /// Retrieves single ECU jobs from the current variants `DiagLayer` and its parent
-    /// references, filtered by the provided predicate. Returns an empty vector if no variant
-    /// is set.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let jobs = ecu_manager.get_single_ecu_jobs_from_variant_and_parent_refs(
-    ///     |job| job.diag_comm().and_then(|dc| dc.short_name()) == Some("ReadSerialNumber"),
-    /// );
-    /// ```
-    fn get_single_ecu_jobs_from_variant_and_parent_refs<F>(
-        &self,
-        service_filter: F,
-    ) -> Vec<datatypes::SingleEcuJob<'_>>
-    where
-        F: Fn(&datatypes::SingleEcuJob) -> bool,
-    {
-        self.variant()
-            .and_then(|v| v.diag_layer().map(|dl| (dl, v.parent_refs())))
-            .map_or(<_>::default(), |(diag_layer, parent_refs)| {
-                Self::get_single_ecu_jobs_from_diag_layer_and_parent_refs(
-                    &(diag_layer.into()),
-                    parent_refs.into_iter().flatten().map(datatypes::ParentRef),
-                    service_filter,
-                )
-            })
-    }
-
-    fn get_service_by_location(
-        &self,
-        location: &CacheLocation,
-    ) -> Option<datatypes::DiagService<'_>> {
-        match location {
-            CacheLocation::Variant(idx) => self
-                .variant()
-                // This is necessary, so we are able to lookup services
-                // _before_ a variant has been found i.e. for variant detection.
-                .or_else(|| self.diag_database.base_variant().ok())
-                .and_then(|v| v.diag_layer())
-                .and_then(|dl| dl.diag_services())
-                .map(|s| s.get(*idx))
-                .map(datatypes::DiagService),
-            CacheLocation::ParentRef(idx) => self
-                .get_variant_parent_ref_services()
-                .and_then(|services| services.get(*idx).cloned()),
-        }
-    }
-
-    /// Recursively resolves parent references and collects their associated `DiagComm` entries.
-    /// Traverses the parent reference hierarchy to gather `DiagComms` from
-    /// inherited `DiagLayers`. Items whose short name appears in a parent references
-    /// `not_inherited_diag_comm_short_names` list are excluded.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let services = EcuManager::get_parent_ref_diag_comms_recursive(
-    ///     parent_refs.into_iter().map(datatypes::ParentRef),
-    ///     |dl| dl.diag_services().map(|s| s.iter().map(datatypes::DiagService).collect()),
-    ///     |service| service.diag_comm().and_then(|dc| dc.short_name()),
-    /// );
-    /// ```
-    fn get_parent_ref_diag_comms_recursive<'a, T>(
-        parent_refs: impl Iterator<Item = impl Into<datatypes::ParentRef<'a>>>,
-        extract: impl Fn(&datatypes::DiagLayer<'a>) -> Option<Vec<T>>,
-        get_name: impl Fn(&T) -> Option<&str>,
-    ) -> Option<Vec<T>> {
-        let all_items: Vec<T> = Self::get_parent_ref_diag_layers_with_refs_recursive(parent_refs)
-            .into_iter()
-            .filter_map(|(parent_ref, diag_layer)| {
-                let not_inherited_names: Vec<&str> = parent_ref
-                    .not_inherited_diag_comm_short_names()
-                    .map_or(<_>::default(), |names| names.iter().collect());
-
-                extract(&diag_layer).map(|items| {
-                    items
-                        .into_iter()
-                        .filter(|item| {
-                            get_name(item).is_none_or(|name| !not_inherited_names.contains(&name))
-                        })
-                        .collect::<Vec<_>>()
-                })
-            })
-            .flatten()
-            .collect();
-
-        if all_items.is_empty() {
-            None
-        } else {
-            Some(all_items)
-        }
-    }
-
-    /// Recursively resolves parent references and collects their single ECU jobs.
-    /// Traverses the parent reference hierarchy to gather jobs from inherited `DiagLayers`.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let jobs = EcuManager::get_parent_ref_jobs_recursive(
-    ///     parent_refs.into_iter().map(datatypes::ParentRef),
-    /// );
-    /// // jobs: Option<Vec<datatypes::SingleEcuJob>>
-    /// ```
-    fn get_parent_ref_jobs_recursive<'a>(
-        parent_refs: impl Iterator<Item = impl Into<datatypes::ParentRef<'a>>>,
-    ) -> Option<Vec<datatypes::SingleEcuJob<'a>>> {
-        Self::get_parent_ref_diag_comms_recursive(
-            parent_refs,
-            |dl| {
-                dl.single_ecu_jobs()
-                    .map(|jobs| jobs.iter().map(datatypes::SingleEcuJob).collect())
-            },
-            |job| job.diag_comm().and_then(|dc| dc.short_name()),
-        )
-    }
-
-    /// Recursively resolves parent references and collects their diagnostic services.
-    /// Traverses the parent reference hierarchy to gather services
-    /// from inherited `DiagLayers`.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let services = EcuManager::get_parent_ref_services_recursive(
-    ///     parent_refs.into_iter().map(datatypes::ParentRef),
-    /// );
-    /// // services: Option<Vec<datatypes::DiagService>>
-    /// ```
-    fn get_parent_ref_services_recursive<'a>(
-        parent_refs: impl Iterator<Item = impl Into<datatypes::ParentRef<'a>>>,
-    ) -> Option<Vec<datatypes::DiagService<'a>>> {
-        Self::get_parent_ref_diag_comms_recursive(
-            parent_refs,
-            |dl| {
-                dl.diag_services()
-                    .map(|s| s.iter().map(datatypes::DiagService).collect())
-            },
-            |service| service.diag_comm().and_then(|dc| dc.short_name()),
-        )
-    }
-
-    /// Retrieves `DiagServices` inherited from the current variants parent references.
-    /// Falls back to the base variant if no variant has been identified yet, which allows
-    /// service lookups during variant detection.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let inherited_services: Option<Vec<datatypes::DiagService>> =
-    ///     ecu_manager.get_variant_parent_ref_services();
-    /// if let Some(services) = inherited_services {
-    ///     for service in &services {
-    ///         println!("{:?}", service.diag_comm().and_then(|dc| dc.short_name()));
-    ///     }
-    /// }
-    /// ```
-    fn get_variant_parent_ref_services(&self) -> Option<Vec<datatypes::DiagService<'_>>> {
-        self.variant()
-            // This is necessary, so we are able to lookup services
-            // _before_ a variant has been found i.e. for variant detection.
-            .or_else(|| self.diag_database.base_variant().ok())
-            .and_then(|v| v.parent_refs())
-            .and_then(|parent_refs| {
-                Self::get_parent_ref_services_recursive(
-                    parent_refs.iter().map(datatypes::ParentRef::from),
-                )
-            })
-    }
-
-    /// Collects all `DiagLayers` from the current variant and its parent references.
-    /// The variants own `DiagLayer` is placed first to give it higher priority in
-    /// subsequent operations, followed by layers resolved recursively from parent references.
-    ///
-    /// When no specific variant has been detected yet, falls back to the base variant so
-    /// that state-chart and service lookups work correctly during variant detection.
-    /// Returns an empty vector only if neither a detected variant nor a base variant
-    /// is available.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let diag_layers: Vec<datatypes::DiagLayer> =
-    ///     ecu_manager.get_diag_layers_from_variant_and_parent_refs();
-    /// // The first element (if present) is the variants own diagnostic layer.
-    /// for layer in &diag_layers {
-    ///     println!("{:?}", layer.short_name());
-    /// }
-    /// ```
-    fn get_diag_layers_from_variant_and_parent_refs(&self) -> Vec<datatypes::DiagLayer<'_>> {
-        // Fall back to the base variant when no specific variant has been detected yet.
-        // This is necessary to allow state-chart and service lookups during variant detection,
-        // consistent with the approach used in `get_variant_parent_ref_services` and
-        // the service lookup functions.
-        let Some(variant) = self
-            .variant()
-            .or_else(|| self.diag_database.base_variant().ok())
-        else {
-            return Vec::new();
-        };
-
-        // Start with the diag layer of the current variant, to give it a higher
-        // prio in later operations
-        variant
-            .diag_layer()
-            .map(datatypes::DiagLayer)
-            .into_iter()
-            .chain(
-                variant
-                    .parent_refs()
-                    .map(|refs| Self::get_parent_ref_diag_layers_recursive(refs.iter()))
-                    .unwrap_or_default(),
-            )
-            .collect()
-    }
-
-    /// Recursively resolves parent references and collects their `DiagLayers`.
-    /// This is a convenience wrapper around [`get_parent_ref_diag_layers_with_refs_recursive`]
-    /// that discards the associated `ParentRef` and returns only the `DiagLayer` values.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let layers: Vec<datatypes::DiagLayer> = EcuManager::get_parent_ref_diag_layers_recursive(
-    ///     parent_refs.iter().map(datatypes::ParentRef),
-    /// );
-    /// ```
-    fn get_parent_ref_diag_layers_recursive<'a>(
-        parent_refs: impl Iterator<Item = impl Into<datatypes::ParentRef<'a>>>,
-    ) -> Vec<datatypes::DiagLayer<'a>> {
-        Self::get_parent_ref_diag_layers_with_refs_recursive(parent_refs)
-            .into_iter()
-            .map(|(_, diag_layer)| diag_layer)
-            .collect()
-    }
-
-    /// Recursively resolves parent references and returns `(ParentRef, DiagLayer)` pairs.
-    /// Uses a stack-based traversal to handle the parent reference hierarchy:
-    /// - **`FunctionalGroup`**: extracts the `DiagLayer` and pushes its nested `ParentRef`s
-    ///   onto the stack for further traversal.
-    /// - **`Variant`**: extracts the `DiagLayer` and pushes its nested `ParentRef`s
-    ///   onto the stack for further traversal.
-    /// - **`Protocol`**: extracts the `DiagLayer` and pushes its nested `ParentRef`
-    ///   items onto the stack for further traversal.
-    /// - **`EcuSharedData`**: extracts the `DiagLayer` (leaf node, no `parent_refs`).
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let pairs: Vec<(datatypes::ParentRef, datatypes::DiagLayer)> =
-    ///     EcuManager::get_parent_ref_diag_layers_with_refs_recursive(
-    ///         parent_refs.iter().map(datatypes::ParentRef),
-    ///     );
-    /// for (parent_ref, diag_layer) in &pairs {
-    ///     println!("ref type: {:?}, layer: {:?}", parent_ref.ref_type(), diag_layer.short_name());
-    /// }
-    /// ```
-    fn get_parent_ref_diag_layers_with_refs_recursive<'a>(
-        parent_refs: impl Iterator<Item = impl Into<datatypes::ParentRef<'a>>>,
-    ) -> Vec<(datatypes::ParentRef<'a>, datatypes::DiagLayer<'a>)> {
-        let mut result = Vec::new();
-        let mut stack: Vec<datatypes::ParentRef<'a>> =
-            parent_refs.into_iter().map(Into::into).collect();
-
-        while let Some(parent_ref) = stack.pop() {
-            match parent_ref.ref_type().try_into() {
-                Ok(datatypes::ParentRefType::FunctionalGroup) => {
-                    if let Some(fg) = parent_ref.ref__as_functional_group() {
-                        if let Some(nested_refs) = fg.parent_refs() {
-                            stack.extend(nested_refs.iter().map(datatypes::ParentRef));
-                        }
-                        if let Some(dl) = fg.diag_layer() {
-                            result.push((parent_ref, datatypes::DiagLayer(dl)));
-                        }
-                    }
-                }
-                Ok(datatypes::ParentRefType::EcuSharedData) => {
-                    if let Some(dl) = parent_ref
-                        .ref__as_ecu_shared_data()
-                        .and_then(|esd| esd.diag_layer())
-                    {
-                        result.push((parent_ref, datatypes::DiagLayer(dl)));
-                    }
-                }
-                Ok(datatypes::ParentRefType::Protocol) => {
-                    if let Some(p) = parent_ref.ref__as_protocol() {
-                        if let Some(nested_refs) = p.parent_refs() {
-                            stack.extend(nested_refs.iter().map(datatypes::ParentRef));
-                        }
-                        if let Some(dl) = p.diag_layer() {
-                            result.push((parent_ref, datatypes::DiagLayer(dl)));
-                        }
-                    }
-                }
-                Ok(datatypes::ParentRefType::Variant) => {
-                    if let Some(v) = parent_ref.ref__as_variant() {
-                        if let Some(nested_refs) = v.parent_refs() {
-                            stack.extend(nested_refs.iter().map(datatypes::ParentRef));
-                        }
-                        if let Some(dl) = v.diag_layer() {
-                            result.push((parent_ref, datatypes::DiagLayer(dl)));
-                        }
-                    }
-                }
-                _ => {
-                    tracing::error!("Unsupported ParentRefType in ECU shared service lookup.");
-                }
-            }
-        }
-
-        result
     }
 
     #[tracing::instrument(skip_all,
@@ -5081,26 +4318,6 @@ impl<S: SecurityPlugin> EcuManager<S> {
             ))
     }
 
-    fn lookup_services_by_sid(
-        &self,
-        service_id: u8,
-    ) -> Result<Vec<datatypes::DiagService<'_>>, DiagServiceError> {
-        let services = self.get_services_from_variant_and_parent_refs(|service| {
-            service
-                .request_id()
-                .is_some_and(|req_id| req_id == service_id)
-        });
-
-        if services.is_empty() {
-            Err(DiagServiceError::NotFound(format!(
-                "No services with SID {service_id:#04X} found in variant, base variant, or ECU \
-                 shared data"
-            )))
-        } else {
-            Ok(services)
-        }
-    }
-
     fn find_dtc_dop_in_params<'a>(
         params: &Vec<datatypes::Parameter<'a>>,
     ) -> Result<Option<datatypes::DtcDop<'a>>, DiagServiceError> {
@@ -5611,136 +4828,43 @@ mod tests {
     use std::vec;
 
     use cda_database::datatypes::{
-        CompuCategory, DataType, DiagCodedTypeVariant, Limit, ResponseType,
-        database_builder::{
-            Addressing, DataFormatParentRefType, DiagClassType, DiagCommParams, DiagLayerParams,
-            DiagServiceParams, DopType, EcuDataBuilder, EcuDataParams, SpecificDOPData,
-            TransmissionMode,
-        },
+        DataType,
+        database_builder::{DiagClassType, DiagCommParams, DiagLayerParams, EcuDataBuilder},
     };
     use cda_interfaces::{EcuManager, Protocol, UDS_ID_RESPONSE_BITMASK};
     use cda_plugin_security::DefaultSecurityPluginData;
-    use flatbuffers::WIPOffset;
     use serde_json::json;
 
     use super::*;
-
-    const SID_PARM_NAME: &str = "sid";
-
-    const TEST_DIAG_LAYER: &str = "TestLayer";
+    use crate::diag_kernel::test_utils::{
+        db_builder::{finish_db, finish_db_with_functional_groups},
+        ecu_manager_builder::{
+            EndOfPduStructureType, SID_PARM_NAME, ServiceSecurityTransition,
+            create_ecu_manager_dlf_sibling_no_byte_pos, create_ecu_manager_env_data_no_wildcard,
+            create_ecu_manager_variant_detection, create_ecu_manager_with_dtc,
+            create_ecu_manager_with_dynamic_length_field_service,
+            create_ecu_manager_with_end_pdu_service, create_ecu_manager_with_env_data_desc,
+            create_ecu_manager_with_env_data_desc_wildcard,
+            create_ecu_manager_with_length_key_request_service,
+            create_ecu_manager_with_mixed_functional_group, create_ecu_manager_with_mux_service,
+            create_ecu_manager_with_mux_service_and_default_case,
+            create_ecu_manager_with_param_length_info_service,
+            create_ecu_manager_with_parameter_metadata,
+            create_ecu_manager_with_phys_const_normal_dop_service,
+            create_ecu_manager_with_phys_const_structure_dop_service,
+            create_ecu_manager_with_preconditions_and_functional_group,
+            create_ecu_manager_with_state_transitions,
+            create_ecu_manager_with_static_field_service, create_ecu_manager_with_struct_service,
+            create_ecu_manager_with_trailing_param_after_param_length_info_service,
+            new_ecu_manager,
+        },
+        mdd_type_builder::{create_sid_only_request, new_diag_comm, new_diag_service},
+    };
 
     macro_rules! skip_sec_plugin {
         () => {{
             let skip_sec_plugin: DynamicPlugin = Box::new(());
             skip_sec_plugin
-        }};
-    }
-
-    /// Helper: finish an `EcuDataBuilder` into a `DiagnosticDatabase` containing
-    /// a single variant with one `DiagLayer` that holds the given diag services.
-    ///
-    /// Delegates to [`EcuDataBuilder::finish_with_single_variant`] with
-    /// hardcoded test defaults for layer name, ECU name, revision and version.
-    macro_rules! finish_db {
-        ($builder:expr, $protocol:expr, $diag_services:expr) => {
-            $builder.finish_with_single_variant(
-                $protocol,
-                $diag_services,
-                TEST_DIAG_LAYER,
-                "TestEcu",
-                "1",
-                "1.0.0",
-            )
-        };
-    }
-
-    /// Helper: build a database with a single variant and functional groups.
-    macro_rules! finish_db_with_functional_groups {
-        ($builder:expr, $protocol:expr, $variant_services:expr, $functional_groups:expr) => {{
-            let cp_ref = $builder.create_com_param_ref(None, None, None, Some($protocol), None);
-            let diag_layer = $builder.create_diag_layer(DiagLayerParams {
-                short_name: TEST_DIAG_LAYER,
-                com_param_refs: Some(vec![cp_ref]),
-                diag_services: {
-                    let services: Vec<_> = $variant_services;
-                    if services.is_empty() {
-                        None
-                    } else {
-                        Some(services)
-                    }
-                },
-                ..Default::default()
-            });
-            let variant = $builder.create_variant(diag_layer, true, None, None);
-            $builder.finish(EcuDataParams {
-                ecu_name: "TestEcu",
-                revision: "1",
-                version: "1.0.0",
-                variants: Some(vec![variant]),
-                functional_groups: Some($functional_groups),
-                ..Default::default()
-            })
-        }};
-    }
-
-    /// Helper: build a `DiagComm` flatbuffer node with test-default fields.
-    macro_rules! new_diag_comm {
-        ($builder:expr, $name:expr, $protocol:expr) => {
-            $builder.create_diag_comm(DiagCommParams {
-                short_name: $name,
-                diag_class_type: DiagClassType::START_COMM,
-                protocols: Some(vec![$protocol]),
-                ..Default::default()
-            })
-        };
-    }
-
-    /// Helper: build a `DiagService` flatbuffer node with test-default fields.
-    macro_rules! new_diag_service {
-        ($builder:expr, $diag_comm:expr, $request:expr, $pos:expr, $neg:expr) => {
-            $builder.create_diag_service(DiagServiceParams {
-                diag_comm: Some($diag_comm),
-                request: Some($request),
-                pos_responses: $pos,
-                neg_responses: $neg,
-                addressing: *Addressing::FUNCTIONAL_OR_PHYSICAL,
-                transmission_mode: *TransmissionMode::SEND_AND_RECEIVE,
-                ..Default::default()
-            })
-        };
-    }
-
-    /// Helper macro: create a CODED-CONST SID parameter at byte position 0.
-    macro_rules! create_sid_param {
-        ($builder:expr, $name:expr, $sid:expr) => {
-            $builder.create_coded_const_param($name, &$sid.to_string(), 0, 0, 8, DataType::UInt32)
-        };
-        ($builder:expr, $sid:expr) => {
-            create_sid_param!($builder, SID_PARM_NAME, $sid)
-        };
-    }
-
-    /// Helper macro: create a request containing only a SID parameter.
-    macro_rules! create_sid_only_request {
-        ($builder:expr, $name:expr, $sid:expr) => {{
-            let sid_param = create_sid_param!($builder, $name, $sid);
-            $builder.create_request(Some(vec![sid_param]), None)
-        }};
-        ($builder:expr, $sid:expr) => {
-            create_sid_only_request!($builder, SID_PARM_NAME, $sid)
-        };
-    }
-
-    /// Helper macro: create a positive response with a SID param and one value param.
-    macro_rules! create_pos_response_with_param {
-        ($builder:expr, $sid:expr, $param_name:expr, $dop:expr, $byte_pos:expr) => {{
-            let sid_param = create_sid_param!($builder, "test_service_pos_sid", $sid);
-            let value_param = $builder.create_value_param($param_name, $dop, $byte_pos, 0);
-            $builder.create_response(
-                ResponseType::Positive,
-                Some(vec![sid_param, value_param]),
-                None,
-            )
         }};
     }
 
@@ -5780,2019 +4904,6 @@ mod tests {
             .convert_from_uds(service, &create_payload(payload_data), true, None)
             .await;
         assert!(response.is_ok(), "Expected convert_from_uds to succeed");
-    }
-
-    #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-    enum EndOfPduStructureType {
-        FixedSize,
-        LeadingLengthDop,
-    }
-
-    fn new_ecu_manager(
-        db: datatypes::DiagnosticDatabase,
-    ) -> super::EcuManager<DefaultSecurityPluginData> {
-        let mut manager = super::EcuManager::new(
-            db,
-            Protocol::default(),
-            &ComParams::default(),
-            DatabaseNamingConvention::default(),
-            super::EcuManagerConfig {
-                type_: EcuManagerType::Ecu,
-                fallback_to_base_variant: true,
-                strict_parameter_validation: false,
-            },
-            &cda_interfaces::FunctionalDescriptionConfig {
-                description_database: "functional_groups".to_owned(),
-                enabled_functional_groups: None,
-                protocol_position:
-                    cda_interfaces::datatypes::DiagnosticServiceAffixPosition::Suffix,
-            },
-        )
-        .expect("Failed to create EcuManager");
-
-        // not using set_variant here, because that would require us to build state charts etc.
-        manager.variant = EcuVariant {
-            name: Some(TEST_DIAG_LAYER.to_owned()),
-            is_base_variant: true,
-            is_fallback: false,
-            state: EcuState::Online,
-            logical_address: 0,
-        };
-        manager.variant_index = Some(0);
-
-        manager
-    }
-
-    fn new_ecu_manager_no_base_fallback(
-        db: datatypes::DiagnosticDatabase,
-    ) -> super::EcuManager<DefaultSecurityPluginData> {
-        super::EcuManager::new(
-            db,
-            Protocol::default(),
-            &ComParams::default(),
-            DatabaseNamingConvention::default(),
-            super::EcuManagerConfig {
-                type_: EcuManagerType::Ecu,
-                fallback_to_base_variant: false,
-                strict_parameter_validation: false,
-            },
-            &cda_interfaces::FunctionalDescriptionConfig {
-                description_database: "functional_groups".to_owned(),
-                enabled_functional_groups: None,
-                protocol_position:
-                    cda_interfaces::datatypes::DiagnosticServiceAffixPosition::Suffix,
-            },
-        )
-        .unwrap()
-    }
-
-    /// Creates an ECU manager whose database contains a functional group named `"MixedGroup"`
-    /// with one `ReadDataByIdentifier` service (`"ReadService"`) and one
-    /// `WriteDataByIdentifier` service (`"WriteService"`).
-    fn create_ecu_manager_with_mixed_functional_group()
-    -> super::EcuManager<DefaultSecurityPluginData> {
-        let mut db_builder = EcuDataBuilder::new();
-        let protocol_name = Protocol::default().to_string();
-        let protocol = db_builder.create_protocol(&protocol_name, None, None, None);
-
-        // Create a READ_DATA_BY_IDENTIFIER service
-        let read_diag_comm = db_builder.create_diag_comm(DiagCommParams {
-            short_name: "ReadService",
-            long_name: Some("Read Service"),
-            semantic: Some("DATA"),
-            protocols: Some(vec![protocol]),
-            ..Default::default()
-        });
-        let read_request =
-            create_sid_only_request!(db_builder, service_ids::READ_DATA_BY_IDENTIFIER);
-        let read_service =
-            new_diag_service!(db_builder, read_diag_comm, read_request, vec![], vec![]);
-
-        // Create a WRITE_DATA_BY_IDENTIFIER service
-        let write_diag_comm = db_builder.create_diag_comm(DiagCommParams {
-            short_name: "WriteService",
-            long_name: Some("Write Service"),
-            semantic: Some("DATA"),
-            protocols: Some(vec![protocol]),
-            ..Default::default()
-        });
-        let write_request =
-            create_sid_only_request!(db_builder, service_ids::WRITE_DATA_BY_IDENTIFIER);
-        let write_service =
-            new_diag_service!(db_builder, write_diag_comm, write_request, vec![], vec![]);
-
-        let fg_diag_layer = db_builder.create_diag_layer(DiagLayerParams {
-            short_name: "MixedGroup",
-            diag_services: Some(vec![read_service, write_service]),
-            ..Default::default()
-        });
-        let fg = db_builder.create_functional_group(fg_diag_layer, None);
-
-        let db = finish_db_with_functional_groups!(db_builder, protocol, vec![], vec![fg]);
-        new_ecu_manager(db)
-    }
-
-    /// Creates an ECU manager with a diagnostic service containing a `DynamicLengthField` DOP.
-    ///
-    /// # Database contents:
-    /// - **Service**: `TestDynamicLengthFieldService` (SID: 0x2E - `WriteDataByIdentifier`)
-    /// - **Request**: Contains a `num_items` parameter (u8) that specifies the count
-    /// - **Response**: Contains a `dynamic_length_field_dop`
-    ///   that repeats a structure based on `num_items`
-    ///   - Each repeated structure contains `item_param` (u16)
-    /// - **DOPs**: `NormalDOP` for `num_items`, `DynamicLengthField` DOP for response
-    // allowed because creation of test data should keep together
-    #[allow(clippy::too_many_lines)]
-    fn create_ecu_manager_with_dynamic_length_field_service() -> (
-        super::EcuManager<DefaultSecurityPluginData>,
-        cda_interfaces::DiagComm,
-        u8,
-    ) {
-        let mut db_builder = EcuDataBuilder::new();
-        let u8_diag_type = db_builder.create_diag_coded_type_standard_length(8, DataType::UInt32);
-        let u16_diag_type = db_builder.create_diag_coded_type_standard_length(16, DataType::UInt32);
-        let protocol_name = Protocol::default().to_string();
-        let protocol = db_builder.create_protocol(&protocol_name, None, None, None);
-        let compu_identical =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-
-        // Create DOPs for structure parameters
-        let num_items_dop =
-            db_builder.create_regular_normal_dop("num_items_dop", u8_diag_type, compu_identical);
-
-        // Create the structure for the repeated item
-        let repeated_struct = {
-            let item_param_dop = db_builder.create_regular_normal_dop(
-                "item_param_dop",
-                u16_diag_type,
-                compu_identical,
-            );
-
-            // Create parameter for the repeated item
-            let item_param = db_builder.create_value_param("item_param", item_param_dop, 0, 0);
-            db_builder.create_structure(Some(vec![item_param]), Some(2), true)
-        };
-
-        let dynamic_length_field_dop = {
-            // Create DynamicLengthField DoP
-            let dynamic_length_field_dop_specific_data = db_builder
-                .create_dynamic_length_specific_dop_data(
-                    1,
-                    0,
-                    0,
-                    num_items_dop,
-                    Some(repeated_struct),
-                )
-                .value_offset();
-
-            db_builder.create_dop(
-                *DopType::REGULAR,
-                Some("dynamic_length_field_dop"),
-                None,
-                *SpecificDOPData::DynamicLengthField,
-                Some(dynamic_length_field_dop_specific_data),
-            )
-        };
-
-        let sid = service_ids::WRITE_DATA_BY_IDENTIFIER;
-        let dc_name = "TestDynamicLengthFieldService";
-        let diag_comm = new_diag_comm!(db_builder, dc_name, protocol);
-
-        let request = {
-            let sid_param = create_sid_param!(db_builder, sid);
-            let request_num_items_param =
-                db_builder.create_value_param("num_items", num_items_dop, 1, 0);
-            db_builder.create_request(Some(vec![sid_param, request_num_items_param]), None)
-        };
-
-        // Build response
-        let pos_response = create_pos_response_with_param!(
-            db_builder,
-            sid,
-            "pos_response_param",
-            dynamic_length_field_dop,
-            1
-        );
-
-        let neg_response = {
-            let nack_param = db_builder.create_coded_const_param(
-                "test_service_nack",
-                &service_ids::NEGATIVE_RESPONSE.to_string(),
-                0,
-                0,
-                8,
-                DataType::UInt32,
-            );
-            let sid_param = db_builder.create_coded_const_param(
-                "test_service_neg_sid",
-                &sid.to_string(),
-                1,
-                0,
-                8,
-                DataType::UInt32,
-            );
-            db_builder.create_response(
-                ResponseType::Negative,
-                Some(vec![nack_param, sid_param]),
-                None,
-            )
-        };
-
-        let diag_service = new_diag_service!(
-            db_builder,
-            diag_comm,
-            request,
-            vec![pos_response],
-            vec![neg_response]
-        );
-
-        let db = finish_db!(db_builder, protocol, vec![diag_service]);
-        (
-            new_ecu_manager(db),
-            cda_interfaces::DiagComm::new(dc_name, DiagCommType::Configurations),
-            sid,
-        )
-    }
-
-    /// Creates an ECU manager with a service
-    /// containing different parameter types for metadata testing.
-    ///
-    /// # Database contents:
-    /// - **Service**: `RDBI_TestService` (SID: 0x22 - `ReadDataByIdentifier`)
-    /// - **Request**: Contains three parameter types:
-    ///   - `sid`: CODED-CONST parameter (value: "34")
-    ///   - `RDBI_DID`: CODED-CONST parameter (value: "0xF190" = 61840)
-    ///   - `data`: VALUE parameter (u16)
-    /// - **Response**: Contains positive response SID
-    ///
-    /// This helper is used to test parameter metadata extraction, including
-    /// distinguishing between CODED-CONST and VALUE parameter types.
-    fn create_ecu_manager_with_parameter_metadata() -> super::EcuManager<DefaultSecurityPluginData>
-    {
-        let mut db_builder = EcuDataBuilder::new();
-        let protocol_name = Protocol::default().to_string();
-        let protocol = db_builder.create_protocol(&protocol_name, None, None, None);
-        let compu_identical =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-        let u16_diag_type = db_builder.create_diag_coded_type_standard_length(16, DataType::UInt32);
-
-        // Create a service with CODED-CONST parameters
-        let sid = service_ids::READ_DATA_BY_IDENTIFIER;
-        let dc_name = "RDBI_TestService";
-
-        // Create DOP for VALUE parameter
-        let value_dop =
-            db_builder.create_regular_normal_dop("value_dop", u16_diag_type, compu_identical);
-
-        let diag_comm = new_diag_comm!(db_builder, dc_name, protocol);
-
-        // Create request with CODED-CONST and VALUE parameters
-        let request = {
-            let sid_param = create_sid_param!(db_builder, sid);
-            let did_param = db_builder.create_coded_const_param(
-                "RDBI_DID",
-                "0xF190",
-                1,
-                0,
-                16,
-                DataType::UInt32,
-            );
-            let value_param = db_builder.create_value_param("data", value_dop, 3, 0);
-            db_builder.create_request(Some(vec![sid_param, did_param, value_param]), None)
-        };
-
-        let pos_response = {
-            let sid_param =
-                create_sid_param!(db_builder, "pos_sid", (sid + UDS_ID_RESPONSE_BITMASK));
-            db_builder.create_response(ResponseType::Positive, Some(vec![sid_param]), None)
-        };
-
-        let diag_service =
-            new_diag_service!(db_builder, diag_comm, request, vec![pos_response], vec![]);
-
-        let db = finish_db!(db_builder, protocol, vec![diag_service]);
-        new_ecu_manager(db)
-    }
-
-    /// Creates an ECU manager with a diagnostic service containing a Structure DOP.
-    ///
-    /// # Database contents:
-    /// - **Service**: `TestStructService` (SID: 0x2E - `WriteDataByIdentifier`)
-    /// - **Request**: Contains a `main_param` that is a Structure DOP
-    ///   - `param1`: u16
-    ///   - `param2`: f32
-    ///   - `param3`: ASCII string (32 bits)
-    /// - **Structure**: 10 bytes total (2 + 4 + 4)
-    /// - **DOPs**: `NormalDOPs` for each parameter, wrapped in a Structure DOP
-    ///
-    /// # Parameters:
-    /// - `struct_byte_pos`: The byte position where the structure starts in the payload
-    // allowed because creation of test data should kept together
-    fn create_ecu_manager_with_struct_service(
-        struct_byte_pos: u32,
-    ) -> (
-        super::EcuManager<DefaultSecurityPluginData>,
-        cda_interfaces::DiagComm,
-        u8,
-        u32,
-    ) {
-        let mut db_builder = EcuDataBuilder::new();
-        let protocol_name = Protocol::default().to_string();
-        let protocol = db_builder.create_protocol(&protocol_name, None, None, None);
-        let compu_identical =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-
-        // Create the structure with parameters
-        let (structure_dop, structure_byte_len) = {
-            let u16_diag_type =
-                db_builder.create_diag_coded_type_standard_length(16, DataType::UInt32);
-            let f32_diag_type =
-                db_builder.create_diag_coded_type_standard_length(32, DataType::Float32);
-            let ascii_diag_type =
-                db_builder.create_diag_coded_type_standard_length(32, DataType::AsciiString);
-
-            // Create DOPs for structure parameters
-            let param1_dop =
-                db_builder.create_regular_normal_dop("param1_dop", u16_diag_type, compu_identical);
-            let param2_dop =
-                db_builder.create_regular_normal_dop("param2_dop", f32_diag_type, compu_identical);
-            let param3_dop = db_builder.create_regular_normal_dop(
-                "param3_dop",
-                ascii_diag_type,
-                compu_identical,
-            );
-
-            // Create parameters for the structure
-            let struct_param1 = db_builder.create_value_param("param1", param1_dop, 0, 0);
-            let struct_param2 = db_builder.create_value_param("param2", param2_dop, 2, 0);
-            let struct_param3 = db_builder.create_value_param("param3", param3_dop, 6, 0);
-
-            let struct_byte_len = 10; // 2 + 4 + 4 bytes
-            let structure = db_builder.create_structure(
-                Some(vec![struct_param1, struct_param2, struct_param3]),
-                Some(struct_byte_len),
-                true,
-            );
-
-            // Wrap the structure in a DOP
-            (
-                db_builder.create_structure_dop("test_structure_dop", structure),
-                struct_byte_len,
-            )
-        };
-
-        let sid = service_ids::WRITE_DATA_BY_IDENTIFIER;
-        let dc_name = "TestStructService";
-        let diag_comm = new_diag_comm!(db_builder, dc_name, protocol);
-
-        let request = {
-            let sid_param = create_sid_param!(db_builder, sid);
-            let main_param =
-                db_builder.create_value_param("main_param", structure_dop, struct_byte_pos, 0);
-            db_builder.create_request(Some(vec![sid_param, main_param]), None)
-        };
-
-        let diag_service = new_diag_service!(db_builder, diag_comm, request, vec![], vec![]);
-
-        let db = finish_db!(db_builder, protocol, vec![diag_service]);
-        (
-            new_ecu_manager(db),
-            cda_interfaces::DiagComm::new(dc_name, DiagCommType::Configurations),
-            sid,
-            structure_byte_len,
-        )
-    }
-
-    /// Creates an ECU manager with a MUX service that includes a default case.
-    ///
-    /// # Database contents:
-    /// - **Service**: `TestMuxService` (SID: `ReadDataByIdentifier` - 0x22)
-    /// - **MUX DOP**: Multiplexer with switch key and cases
-    ///   - **Switch key**: u16 at byte position 0
-    ///   - **Case 1** (range 1-10): Contains f32 and u8 parameters
-    ///   - **Case 2** (range 11-600): Contains i16 and ASCII string parameters
-    ///   - **Case 3** (string "test"): No structure
-    ///   - **Default case**: Contains `default_structure_param_1` (u8)
-    /// - **Request and Response**: Both contain the MUX parameter
-    fn create_ecu_manager_with_mux_service_and_default_case() -> (
-        super::EcuManager<DefaultSecurityPluginData>,
-        cda_interfaces::DiagComm,
-        u8,
-    ) {
-        let mut db_builder = EcuDataBuilder::new();
-        let u8_diag_type = db_builder.create_diag_coded_type_standard_length(8, DataType::UInt32);
-        let compu_identical =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-
-        // Create DOP for default structure parameter
-        let default_structure_param_1 = {
-            let default_structure_param_1_dop = db_builder.create_regular_normal_dop(
-                "default_structure_param_1_dop",
-                u8_diag_type,
-                compu_identical,
-            );
-            db_builder.create_value_param(
-                "default_structure_param_1",
-                default_structure_param_1_dop,
-                0,
-                0,
-            )
-        };
-
-        // Create default structure
-        let default_structure =
-            db_builder.create_structure(Some(vec![default_structure_param_1]), Some(1), true);
-        let default_case = db_builder.create_default_case("default_case", Some(default_structure));
-
-        create_ecu_manager_with_mux_service(Some(db_builder), None, Some(default_case))
-    }
-
-    /// Creates an ECU manager with a MUX (multiplexer) service.
-    ///
-    /// # Database contents:
-    /// - **Service**: `TestMuxService` (SID: `ReadDataByIdentifier` - 0x22)
-    /// - **MUX DOP**: Multiplexer with configurable switch key and cases
-    ///   - **Switch key**: Configurable via parameter, defaults to u16 at byte position 0
-    ///   - **Case 1** (range 1-10):
-    ///     - `mux_1_case_1_param_1`: f32 at byte 0
-    ///     - `mux_1_case_1_param_2`: u8 at byte 4
-    ///   - **Case 2** (range 11-600):
-    ///     - `mux_1_case_2_param_1`: i16 at byte 1
-    ///     - `mux_1_case_2_param_2`: ASCII string (32 bits) at byte 4
-    ///   - **Case 3** (string "test"): No structure
-    ///   - **Default case**: Optional, configurable via parameter
-    /// - **Request and Response**: Both contain the MUX parameter at byte position 2
-    ///
-    /// # Parameters:
-    /// - `db_builder`: Optional pre-configured builder (creates new if None)
-    /// - `switch_key`: Optional custom switch key (creates default u16 if None)
-    /// - `default_case`: Optional default case for unmatched switch values
-    // allowed because creation of test data should kept together
-    fn create_ecu_manager_with_mux_service(
-        db_builder: Option<EcuDataBuilder>,
-        switch_key: Option<WIPOffset<datatypes::database_builder::SwitchKey>>,
-        default_case: Option<WIPOffset<datatypes::database_builder::DefaultCase>>,
-    ) -> (
-        super::EcuManager<DefaultSecurityPluginData>,
-        cda_interfaces::DiagComm,
-        u8,
-    ) {
-        let mut db_builder = db_builder.unwrap_or_default();
-
-        let u8_diag_type = db_builder.create_diag_coded_type_standard_length(8, DataType::UInt32);
-        let u16_diag_type = db_builder.create_diag_coded_type_standard_length(16, DataType::UInt32);
-        let i16_diag_type = db_builder.create_diag_coded_type_standard_length(16, DataType::Int32);
-        let f32_diag_type =
-            db_builder.create_diag_coded_type_standard_length(32, DataType::Float32);
-        let ascii_diag_type =
-            db_builder.create_diag_coded_type_standard_length(32, DataType::AsciiString);
-        let protocol_name = Protocol::default().to_string();
-        let protocol = db_builder.create_protocol(&protocol_name, None, None, None);
-        let compu_identical =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-
-        // Create DOPs for case 1 parameters
-        let mux_1_case_1_params_dop_1 = db_builder.create_regular_normal_dop(
-            "mux_1_case_1_params_dop_1",
-            f32_diag_type,
-            compu_identical,
-        );
-        let mux_1_case_1_params_dop_2 = db_builder.create_regular_normal_dop(
-            "mux_1_case_1_params_dop_2",
-            u8_diag_type,
-            compu_identical,
-        );
-
-        // Create DOPs for case 2 parameters
-        let mux_1_case_2_params_dop_1 = db_builder.create_regular_normal_dop(
-            "mux_1_case_2_params_dop_1",
-            i16_diag_type,
-            compu_identical,
-        );
-        let mux_1_case_2_params_dop_2 = db_builder.create_regular_normal_dop(
-            "mux_1_case_2_params_dop_2",
-            ascii_diag_type,
-            compu_identical,
-        );
-
-        // Create parameters for case 1
-        let mux_1_case_1_param_1 =
-            db_builder.create_value_param("mux_1_case_1_param_1", mux_1_case_1_params_dop_1, 0, 0);
-        let mux_1_case_1_param_2 =
-            db_builder.create_value_param("mux_1_case_1_param_2", mux_1_case_1_params_dop_2, 4, 0);
-
-        // Create parameters for case 2
-        let mux_1_case_2_param_1 =
-            db_builder.create_value_param("mux_1_case_2_param_1", mux_1_case_2_params_dop_1, 1, 0);
-        let mux_1_case_2_param_2 =
-            db_builder.create_value_param("mux_1_case_2_param_2", mux_1_case_2_params_dop_2, 4, 0);
-
-        // Create structures
-        let mux_1_case_1_structure = db_builder.create_structure(
-            Some(vec![mux_1_case_1_param_1, mux_1_case_1_param_2]),
-            Some(7),
-            true,
-        );
-
-        let mux_1_case_2_structure = db_builder.create_structure(
-            Some(vec![mux_1_case_2_param_1, mux_1_case_2_param_2]),
-            Some(7),
-            true,
-        );
-
-        // Create cases using the new helper method
-        let mux_1_case_1 = db_builder.create_case(
-            "mux_1_case_1",
-            Some(Limit {
-                value: 1.0.to_string(),
-                interval_type: datatypes::IntervalType::Infinite,
-            }),
-            Some(Limit {
-                value: 10.0.to_string(),
-                interval_type: datatypes::IntervalType::Infinite,
-            }),
-            Some(mux_1_case_1_structure),
-        );
-
-        let mux_1_case_2 = db_builder.create_case(
-            "mux_1_case_2",
-            Some(Limit {
-                value: 11.0.to_string(),
-                interval_type: datatypes::IntervalType::Infinite,
-            }),
-            Some(Limit {
-                value: 600.0.to_string(),
-                interval_type: datatypes::IntervalType::Infinite,
-            }),
-            Some(mux_1_case_2_structure),
-        );
-
-        let mux_1_case_3 = db_builder.create_case(
-            "mux_1_case_3",
-            Some(Limit {
-                value: "test".to_owned(),
-                interval_type: datatypes::IntervalType::Infinite,
-            }),
-            None,
-            None,
-        );
-
-        // Create switch key if not provided
-        let mux_1_switch_key = switch_key.unwrap_or_else(|| {
-            let switch_key_dop = db_builder.create_regular_normal_dop(
-                "switch_key_dop",
-                u16_diag_type,
-                compu_identical,
-            );
-            db_builder.create_switch_key(0, Some(0), Some(switch_key_dop))
-        });
-
-        let cases = vec![mux_1_case_1, mux_1_case_2, mux_1_case_3];
-
-        // Create mux DOP specific data
-        let mux_dop = db_builder.create_mux_dop(
-            "mux_dop",
-            2,
-            Some(mux_1_switch_key),
-            default_case,
-            Some(cases),
-            true,
-        );
-
-        let sid = service_ids::READ_DATA_BY_IDENTIFIER;
-        let dc_name = "TestMuxService";
-        let diag_comm = new_diag_comm!(db_builder, dc_name, protocol);
-
-        // Create request with mux parameter
-        let request = {
-            let sid_param = create_sid_param!(db_builder, sid);
-            let mux_param = db_builder.create_value_param("mux_1_param", mux_dop, 2, 0);
-            db_builder.create_request(Some(vec![sid_param, mux_param]), None)
-        };
-
-        // Create response with mux parameter
-        let pos_response =
-            create_pos_response_with_param!(db_builder, sid, "mux_1_param", mux_dop, 2);
-
-        let diag_service =
-            new_diag_service!(db_builder, diag_comm, request, vec![pos_response], vec![]);
-
-        let db = finish_db!(db_builder, protocol, vec![diag_service]);
-        (
-            new_ecu_manager(db),
-            cda_interfaces::DiagComm::new(dc_name, DiagCommType::Data),
-            sid,
-        )
-    }
-
-    /// Creates an ECU manager with an `EndOfPdu` service for variable-length repeated structures.
-    ///
-    /// # Database contents:
-    /// - **Service**: `TestEndOfPduService` (SID: `ReadDataByIdentifier` - 0x22)
-    /// - **Request**: Simple request with only SID
-    /// - **Response**: Contains an `end_pdu_param` with `EndOfPdu` DOP
-    ///   - Repeats a structure until end of payload
-    ///   - **Structure type** (configurable):
-    ///     - **`FixedSize`**: Each item is 3 bytes
-    ///       - `item_param1`: u8 at byte 0
-    ///       - `item_param2`: u16 at byte 1
-    ///     - **`LeadingLengthDop`**: Variable-size items with 8-bit length prefix
-    ///       - `data`: `ByteField` with leading length info
-    ///   - **Constraints**:
-    ///   - `min_items`: Minimum number of items required
-    ///   - `max_items`: Optional maximum number of items allowed
-    ///
-    /// # Parameters:
-    /// - `min_items`: Minimum number of structures required
-    /// - `max_items`: Optional maximum number of structures
-    /// - `structure_type`: Whether structures are fixed-size or have leading length
-    // allowed because creation of test data should kept together
-    fn create_ecu_manager_with_end_pdu_service(
-        min_items: u32,
-        max_items: Option<u32>,
-        structure_type: EndOfPduStructureType,
-    ) -> (
-        super::EcuManager<DefaultSecurityPluginData>,
-        cda_interfaces::DiagComm,
-        u8,
-    ) {
-        let mut db_builder = EcuDataBuilder::new();
-        let u8_diag_type = db_builder.create_diag_coded_type_standard_length(8, DataType::UInt32);
-        let u16_diag_type = db_builder.create_diag_coded_type_standard_length(16, DataType::UInt32);
-        let protocol_name = Protocol::default().to_string();
-        let protocol = db_builder.create_protocol(&protocol_name, None, None, None);
-        let compu_identical =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-
-        // Create the structure that will be repeated in EndOfPdu
-        let item_structure = match structure_type {
-            EndOfPduStructureType::LeadingLengthDop => {
-                // Create structure with leading length DOP
-                // The DiagCodedType with LeadingLengthInfo handles the length prefix automatically
-                let leading_length_diag_type = db_builder.create_diag_coded_type(
-                    None,
-                    DataType::ByteField,
-                    true,
-                    DiagCodedTypeVariant::LeadingLengthInfo(8),
-                );
-
-                let data_dop = db_builder.create_regular_normal_dop(
-                    "data_dop",
-                    leading_length_diag_type,
-                    compu_identical,
-                );
-
-                let data_param = db_builder.create_value_param("data", data_dop, 0, 0);
-
-                // Create structure with just the data parameter
-                db_builder.create_structure(Some(vec![data_param]), None, true)
-            }
-            EndOfPduStructureType::FixedSize => {
-                // Create fixed-size structure
-                let item_param1_dop = db_builder.create_regular_normal_dop(
-                    "item_param1_dop",
-                    u8_diag_type,
-                    compu_identical,
-                );
-                let item_param2_dop = db_builder.create_regular_normal_dop(
-                    "item_param2_dop",
-                    u16_diag_type,
-                    compu_identical,
-                );
-
-                // Create parameters for the repeating structure
-                let item_param1 =
-                    db_builder.create_value_param("item_param1", item_param1_dop, 0, 0);
-                let item_param2 =
-                    db_builder.create_value_param("item_param2", item_param2_dop, 1, 0);
-
-                // Create the basic structure that will be repeated
-                db_builder.create_structure(
-                    Some(vec![item_param1, item_param2]),
-                    Some(3), // byte_size: 1 byte + 2 bytes = 3 bytes per item
-                    true,
-                )
-            }
-        };
-
-        // Create EndOfPdu DOP using the new helper method
-        let end_pdu_dop =
-            db_builder.create_end_of_pdu_field_dop(min_items, max_items, Some(item_structure));
-
-        let sid = service_ids::READ_DATA_BY_IDENTIFIER;
-        let dc_name = "TestEndOfPduService";
-        let diag_comm = new_diag_comm!(db_builder, dc_name, protocol);
-
-        // Create request
-        let request = create_sid_only_request!(db_builder, sid);
-
-        // Create response with EndOfPdu parameter
-        let pos_response =
-            create_pos_response_with_param!(db_builder, sid, "end_pdu_param", end_pdu_dop, 1);
-
-        let diag_service =
-            new_diag_service!(db_builder, diag_comm, request, vec![pos_response], vec![]);
-
-        let db = finish_db!(db_builder, protocol, vec![diag_service]);
-        (
-            new_ecu_manager(db),
-            cda_interfaces::DiagComm::new(dc_name, DiagCommType::Data),
-            sid,
-        )
-    }
-
-    /// Creates an ECU manager with a DTC (Diagnostic Trouble Code) service.
-    ///
-    /// # Database contents:
-    /// - **Service**: `TestDtcService` (SID: 0x19 - `ReadDTCInformation`)
-    /// - **Request**: Simple request with only SID
-    /// - **Response**: Contains a `dtc_param` (u32 DTC code)
-    /// - **DTC**: Single DTC definition
-    ///   - **Code**: 0xDEADBEEF (32-bit)
-    ///   - **Display code**: "P1234" (OBD-II format)
-    ///   - **Fault name**: "`TestFault`"
-    ///   - **Severity**: 2
-    /// - **DOP**: DTC DOP with 32-bit coded type
-    fn create_ecu_manager_with_dtc() -> (
-        super::EcuManager<DefaultSecurityPluginData>,
-        cda_interfaces::DiagComm,
-        u8,
-        u32,
-    ) {
-        let mut db_builder = EcuDataBuilder::new();
-        let u32_diag_type = db_builder.create_diag_coded_type_standard_length(32, DataType::UInt32);
-        let protocol_name = Protocol::default().to_string();
-        let protocol = db_builder.create_protocol(&protocol_name, None, None, None);
-        let compu_identical =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-
-        let dtc_code = 0xDEAD_BEEF;
-        let dtc = db_builder.create_dtc(dtc_code, Some("P1234"), Some("TestFault"), 2);
-
-        let dtc_dop =
-            db_builder.create_dtc_dop(u32_diag_type, Some(vec![dtc]), Some(compu_identical));
-
-        let sid = service_ids::READ_DTC_INFORMATION;
-        let dc_name = "TestDtcService";
-        let diag_comm = new_diag_comm!(db_builder, dc_name, protocol);
-
-        // Create request
-        let request = create_sid_only_request!(db_builder, sid);
-
-        // Create response with DTC parameter
-        let pos_response =
-            create_pos_response_with_param!(db_builder, sid, "dtc_param", dtc_dop, 1);
-
-        let diag_service =
-            new_diag_service!(db_builder, diag_comm, request, vec![pos_response], vec![]);
-
-        let db = finish_db!(db_builder, protocol, vec![diag_service]);
-        (
-            new_ecu_manager(db),
-            cda_interfaces::DiagComm::new(dc_name, DiagCommType::Faults),
-            sid,
-            dtc_code,
-        )
-    }
-
-    /// Creates an ECU manager with a service that has a DTC param followed by an ENV-DATA-DESC
-    /// param. The ENV-DATA-DESC selects the right ENV-DATA based on the DTC code and decodes
-    /// a single `temperature` (u8) parameter from it.
-    fn create_ecu_manager_with_env_data_desc() -> (
-        super::EcuManager<DefaultSecurityPluginData>,
-        cda_interfaces::DiagComm,
-        u8,
-        u32,
-    ) {
-        let mut db_builder = EcuDataBuilder::new();
-        let u32_diag_type = db_builder.create_diag_coded_type_standard_length(32, DataType::UInt32);
-        let u8_diag_type = db_builder.create_diag_coded_type_standard_length(8, DataType::UInt32);
-        let default_protocol = Protocol::default();
-        let protocol = db_builder.create_protocol(default_protocol.str(), None, None, None);
-        let compu_identical =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-
-        let dtc_code: u32 = 0x0001_0002;
-        let dtc = db_builder.create_dtc(dtc_code, Some("P0001"), Some("TestEnvFault"), 1);
-        let dtc_dop =
-            db_builder.create_dtc_dop(u32_diag_type, Some(vec![dtc]), Some(compu_identical));
-
-        // ENV-DATA for dtc_code: contains a single u8 param "temperature"
-        let compu_identical2 =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-        let temp_dop =
-            db_builder.create_regular_normal_dop("temp_dop", u8_diag_type, compu_identical2);
-        let temp_param = db_builder.create_value_param("temperature", temp_dop, 0, 0);
-        let env_data_dop = db_builder.create_env_data_dop(&[dtc_code], &[temp_param]);
-
-        // ENV-DATA-DESC: selector is "dtc_param"
-        let env_data_desc_dop =
-            db_builder.create_env_data_desc_dop("env_snapshot", "dtc_param", &[env_data_dop]);
-
-        let sid = service_ids::READ_DTC_INFORMATION;
-        let dc_name = "TestEnvDataDescService";
-        let diag_comm = new_diag_comm!(db_builder, dc_name, protocol);
-
-        let request = create_sid_only_request!(db_builder, sid);
-
-        let sid_param = create_sid_param!(db_builder, "test_service_pos_sid", sid);
-        let dtc_value_param = db_builder.create_value_param("dtc_param", dtc_dop, 1, 0);
-        let env_data_value_param =
-            db_builder.create_value_param("env_snapshot", env_data_desc_dop, 5, 0);
-        let pos_response = db_builder.create_response(
-            ResponseType::Positive,
-            Some(vec![sid_param, dtc_value_param, env_data_value_param]),
-            None,
-        );
-
-        let diag_service =
-            new_diag_service!(db_builder, diag_comm, request, vec![pos_response], vec![]);
-
-        let db = finish_db!(db_builder, protocol, vec![diag_service]);
-        (
-            new_ecu_manager(db),
-            cda_interfaces::DiagComm::new(dc_name, DiagCommType::Faults),
-            sid,
-            dtc_code,
-        )
-    }
-
-    // StaticField DOP
-
-    fn create_ecu_manager_with_static_field_service() -> (
-        super::EcuManager<DefaultSecurityPluginData>,
-        cda_interfaces::DiagComm,
-        u8,
-    ) {
-        let mut db_builder = EcuDataBuilder::new();
-        let u16_diag_type = db_builder.create_diag_coded_type_standard_length(16, DataType::UInt32);
-        let compu_identical =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-        let default_protocol = Protocol::default();
-        let protocol = db_builder.create_protocol(default_protocol.str(), None, None, None);
-
-        let item_dop =
-            db_builder.create_regular_normal_dop("item_dop", u16_diag_type, compu_identical);
-        let item_param = db_builder.create_value_param("item_val", item_dop, 0, 0);
-        let item_structure = db_builder.create_structure(Some(vec![item_param]), Some(2), true);
-
-        // 3 items, each 2 bytes wide
-        let static_field_dop = db_builder.create_static_field_dop(3, 2, item_structure);
-
-        let sid = service_ids::READ_DATA_BY_IDENTIFIER;
-        let dc_name = "TestStaticFieldService";
-        let diag_comm = new_diag_comm!(db_builder, dc_name, protocol);
-        let request = create_sid_only_request!(db_builder, sid);
-        let pos_response =
-            create_pos_response_with_param!(db_builder, sid, "items", static_field_dop, 1);
-        let diag_service =
-            new_diag_service!(db_builder, diag_comm, request, vec![pos_response], vec![]);
-        let db = finish_db!(db_builder, protocol, vec![diag_service]);
-        (
-            new_ecu_manager(db),
-            cda_interfaces::DiagComm::new(dc_name, DiagCommType::Data),
-            sid,
-        )
-    }
-
-    // EnvDataDesc wildcard fallback
-
-    fn create_ecu_manager_with_env_data_desc_wildcard() -> (
-        super::EcuManager<DefaultSecurityPluginData>,
-        cda_interfaces::DiagComm,
-        u8,
-        u32, // specific_dtc
-        u32, // other_dtc (triggers wildcard)
-    ) {
-        let mut db_builder = EcuDataBuilder::new();
-        let u32_diag_type = db_builder.create_diag_coded_type_standard_length(32, DataType::UInt32);
-        let u8_diag_type = db_builder.create_diag_coded_type_standard_length(8, DataType::UInt32);
-        let compu_identical =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-        let compu_identical2 =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-        let compu_identical3 =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-        let default_protocol = Protocol::default();
-        let protocol = db_builder.create_protocol(default_protocol.str(), None, None, None);
-
-        let specific_dtc: u32 = 0x0001_0002;
-        let other_dtc: u32 = 0x0003_0004;
-
-        let dtc1 = db_builder.create_dtc(specific_dtc, Some("P0001"), Some("SpecificFault"), 1);
-        let dtc2 = db_builder.create_dtc(other_dtc, Some("P0002"), Some("OtherFault"), 2);
-        let dtc_dop =
-            db_builder.create_dtc_dop(u32_diag_type, Some(vec![dtc1, dtc2]), Some(compu_identical));
-
-        let temp_dop =
-            db_builder.create_regular_normal_dop("temp_dop", u8_diag_type, compu_identical2);
-        let humidity_dop =
-            db_builder.create_regular_normal_dop("humidity_dop", u8_diag_type, compu_identical3);
-
-        // Specific env_data matches only specific_dtc -> reports "temperature"
-        let temp_param = db_builder.create_value_param("temperature", temp_dop, 0, 0);
-        let specific_env_data_dop = db_builder.create_env_data_dop(&[specific_dtc], &[temp_param]);
-
-        // Wildcard env_data (empty dtc_values) -> reports "humidity"
-        let humidity_param = db_builder.create_value_param("humidity", humidity_dop, 0, 0);
-        let wildcard_env_data_dop = db_builder.create_env_data_dop(&[], &[humidity_param]);
-
-        let env_data_desc_dop = db_builder.create_env_data_desc_dop(
-            "env_snapshot",
-            "dtc_param",
-            &[specific_env_data_dop, wildcard_env_data_dop],
-        );
-
-        let sid = service_ids::READ_DTC_INFORMATION;
-        let dc_name = "TestEnvDataDescWildcardService";
-        let diag_comm = new_diag_comm!(db_builder, dc_name, protocol);
-        let request = create_sid_only_request!(db_builder, sid);
-
-        let sid_param = create_sid_param!(db_builder, "test_service_pos_sid", sid);
-        let dtc_value_param = db_builder.create_value_param("dtc_param", dtc_dop, 1, 0);
-        let env_data_value_param =
-            db_builder.create_value_param("env_snapshot", env_data_desc_dop, 5, 0);
-        let pos_response = db_builder.create_response(
-            ResponseType::Positive,
-            Some(vec![sid_param, dtc_value_param, env_data_value_param]),
-            None,
-        );
-
-        let diag_service =
-            new_diag_service!(db_builder, diag_comm, request, vec![pos_response], vec![]);
-        let db = finish_db!(db_builder, protocol, vec![diag_service]);
-        (
-            new_ecu_manager(db),
-            cda_interfaces::DiagComm::new(dc_name, DiagCommType::Faults),
-            sid,
-            specific_dtc,
-            other_dtc,
-        )
-    }
-
-    fn create_ecu_manager_env_data_no_wildcard() -> (
-        super::EcuManager<DefaultSecurityPluginData>,
-        cda_interfaces::DiagComm,
-        u8,
-        u32, // dtc_in_db (used in payload; no matching env_data)
-    ) {
-        let mut db_builder = EcuDataBuilder::new();
-        let u32_diag_type = db_builder.create_diag_coded_type_standard_length(32, DataType::UInt32);
-        let u8_diag_type = db_builder.create_diag_coded_type_standard_length(8, DataType::UInt32);
-        let compu_id =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-        let compu_id2 =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-        let default_protocol = Protocol::default();
-        let protocol = db_builder.create_protocol(default_protocol.str(), None, None, None);
-
-        let dtc_in_db: u32 = 0x0001_AAAA;
-        let dtc = db_builder.create_dtc(dtc_in_db, Some("P9999"), Some("SomeFault"), 1);
-        let dtc_dop = db_builder.create_dtc_dop(u32_diag_type, Some(vec![dtc]), Some(compu_id));
-
-        // env_data only matches a *different* DTC code; no wildcard
-        let temp_dop = db_builder.create_regular_normal_dop("temp_dop", u8_diag_type, compu_id2);
-        let temp_param = db_builder.create_value_param("temperature", temp_dop, 0, 0);
-        let env_data_dop = db_builder.create_env_data_dop(&[0x0000_1111], &[temp_param]);
-        let env_data_desc_dop =
-            db_builder.create_env_data_desc_dop("env_snapshot", "dtc_param", &[env_data_dop]);
-
-        let sid = service_ids::READ_DTC_INFORMATION;
-        let dc_name = "TestEnvDataNoWildcard";
-        let diag_comm = new_diag_comm!(db_builder, dc_name, protocol);
-        let request = create_sid_only_request!(db_builder, sid);
-        let sid_param = create_sid_param!(db_builder, "test_service_pos_sid", sid);
-        let dtc_value_param = db_builder.create_value_param("dtc_param", dtc_dop, 1, 0);
-        let env_data_value_param =
-            db_builder.create_value_param("env_snapshot", env_data_desc_dop, 5, 0);
-        let pos_response = db_builder.create_response(
-            ResponseType::Positive,
-            Some(vec![sid_param, dtc_value_param, env_data_value_param]),
-            None,
-        );
-        let diag_service =
-            new_diag_service!(db_builder, diag_comm, request, vec![pos_response], vec![]);
-        let db = finish_db!(db_builder, protocol, vec![diag_service]);
-        (
-            new_ecu_manager(db),
-            cda_interfaces::DiagComm::new(dc_name, DiagCommType::Faults),
-            sid,
-            dtc_in_db,
-        )
-    }
-
-    // DynamicLengthField: sibling param before DLF with no byte_position
-
-    fn create_ecu_manager_dlf_sibling_no_byte_pos() -> (
-        super::EcuManager<DefaultSecurityPluginData>,
-        cda_interfaces::DiagComm,
-        u8,
-    ) {
-        let mut db_builder = EcuDataBuilder::new();
-        let u8_diag_type = db_builder.create_diag_coded_type_standard_length(8, DataType::UInt32);
-        let u16_diag_type = db_builder.create_diag_coded_type_standard_length(16, DataType::UInt32);
-        let compu_identical =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-        let compu_identical2 =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-        let compu_identical3 =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-        let default_protocol = Protocol::default();
-        let protocol = db_builder.create_protocol(default_protocol.str(), None, None, None);
-
-        let sibling_dop =
-            db_builder.create_regular_normal_dop("sibling_dop", u8_diag_type, compu_identical);
-        let num_items_dop =
-            db_builder.create_regular_normal_dop("num_items_dop", u8_diag_type, compu_identical2);
-
-        let item_dop =
-            db_builder.create_regular_normal_dop("item_dop", u16_diag_type, compu_identical3);
-        let item_param = db_builder.create_value_param("item_val", item_dop, 0, 0);
-        let repeated_struct = db_builder.create_structure(Some(vec![item_param]), Some(2), true);
-
-        // offset=3: items start at param_abs_byte_pos + 3 (= byte 0 + 3 = byte 3 absolute)
-        // number_of_items_byte_pos=2:
-        // count at param_abs_byte_pos + 2 (= byte 0 + 2 = byte 2 absolute)
-        let dlf_specific = db_builder
-            .create_dynamic_length_specific_dop_data(3, 2, 0, num_items_dop, Some(repeated_struct))
-            .value_offset();
-        let dlf_dop = db_builder.create_dop(
-            *DopType::REGULAR,
-            Some("dlf_dop"),
-            None,
-            *SpecificDOPData::DynamicLengthField,
-            Some(dlf_specific),
-        );
-
-        let sid = service_ids::WRITE_DATA_BY_IDENTIFIER;
-        let dc_name = "TestDlfSiblingNoBytePosService";
-        let diag_comm = new_diag_comm!(db_builder, dc_name, protocol);
-        let request = create_sid_only_request!(db_builder, sid);
-
-        let pos_response = {
-            let sid_param = create_sid_param!(db_builder, "test_service_pos_sid", sid);
-            // sibling at explicit byte_pos=1, advances last_read_byte_pos to 2 after decoding
-            let sibling_param = db_builder.create_value_param("sibling_val", sibling_dop, 1, 0);
-            // DLF with NO explicit byte_position - must use base_offset=0, not last_read_byte_pos
-            let dlf_param = db_builder.create_value_param_no_byte_pos("dlf_items", dlf_dop);
-            db_builder.create_response(
-                ResponseType::Positive,
-                Some(vec![sid_param, sibling_param, dlf_param]),
-                None,
-            )
-        };
-
-        let diag_service =
-            new_diag_service!(db_builder, diag_comm, request, vec![pos_response], vec![]);
-        let db = finish_db!(db_builder, protocol, vec![diag_service]);
-        (
-            new_ecu_manager(db),
-            cda_interfaces::DiagComm::new(dc_name, DiagCommType::Configurations),
-            sid,
-        )
-    }
-    /// Creates an ECU manager configured for variant detection testing.
-    ///
-    /// # Database contents:
-    /// - **Service**: `ReadVariantData`
-    ///   (SID: `ReadDataByIdentifier` - 0x22, `VARIANT_IDENTIFICATION` class)
-    /// - **Request**: Simple request with SID
-    /// - **Response**: Contains `variant_code` parameter (u8)
-    /// - **Variants**: Two variants with pattern matching
-    ///   - **`BaseVariant`** (`is_base=true`):
-    ///     - Pattern matches when `variant_code` = 0
-    ///     - Contains variant detection service
-    ///     - State charts: Session (`DefaultSession`), Security (Locked)
-    ///   - **`SpecificVariant`** (`is_base=false`):
-    ///     - Pattern matches when `variant_code` = 1
-    ///     - Contains variant detection service
-    ///     - State charts: Session (`DefaultSession`), Security (Locked)
-    /// - **ECU name**: "`VariantDetectionEcu`"
-    #[allow(clippy::too_many_lines)] // must be kept together
-    fn create_ecu_manager_variant_detection(
-        fallback_to_base: bool,
-    ) -> super::EcuManager<DefaultSecurityPluginData> {
-        let mut db_builder = EcuDataBuilder::new();
-        let protocol_name = Protocol::default().to_string();
-        let protocol = db_builder.create_protocol(&protocol_name, None, None, None);
-        let cp_ref = db_builder.create_com_param_ref(None, None, None, Some(protocol), None);
-
-        let u8_diag_type = db_builder.create_diag_coded_type_standard_length(8, DataType::UInt32);
-        let compu_identical =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-
-        // Create a variant detection service
-        let vd_service_sid = service_ids::READ_DATA_BY_IDENTIFIER;
-        let vd_service_name = "ReadVariantData";
-
-        // Create DOP for variant code response parameter
-        let variant_code_dop =
-            db_builder.create_regular_normal_dop("variant_code_dop", u8_diag_type, compu_identical);
-
-        // Create the diagnostic communication
-        let vd_diag_comm = db_builder.create_diag_comm(DiagCommParams {
-            short_name: vd_service_name,
-            diag_class_type: DiagClassType::VARIANT_IDENTIFICATION,
-            ..Default::default()
-        });
-
-        let vd_request = create_sid_only_request!(db_builder, "vd_service_sid", vd_service_sid);
-
-        let vd_pos_response = {
-            let sid_param = create_sid_param!(
-                db_builder,
-                "vd_service_pos_sid",
-                (vd_service_sid + UDS_ID_RESPONSE_BITMASK)
-            );
-            let variant_param =
-                db_builder.create_value_param("variant_code", variant_code_dop, 1, 0);
-            db_builder.create_response(
-                ResponseType::Positive,
-                Some(vec![sid_param, variant_param]),
-                None,
-            )
-        };
-
-        let vd_diag_service = new_diag_service!(
-            db_builder,
-            vd_diag_comm,
-            vd_request,
-            vec![vd_pos_response],
-            vec![]
-        );
-
-        // Create state charts for session and security
-        let session_state_chart = {
-            let default_session_name = "DefaultSession";
-            let default_session_state = db_builder.create_state(default_session_name, None);
-
-            db_builder.create_state_chart(
-                "Session",
-                Some(semantics::SESSION),
-                None,
-                Some(default_session_name),
-                Some(vec![default_session_state]),
-            )
-        };
-
-        let security_state_chart = {
-            let default_security_name = "Locked";
-            let default_security_state = db_builder.create_state(default_security_name, None);
-
-            db_builder.create_state_chart(
-                "SecurityAccess",
-                Some(semantics::SECURITY),
-                None,
-                Some(default_security_name),
-                Some(vec![default_security_state]),
-            )
-        };
-
-        let sid_param = create_sid_param!(
-            db_builder,
-            "vd_service_pos_sid_base",
-            (vd_service_sid + UDS_ID_RESPONSE_BITMASK)
-        );
-
-        let variant_param = db_builder.create_value_param("variant_code", variant_code_dop, 1, 0);
-
-        let pos_response_variant_code = {
-            db_builder.create_response(
-                ResponseType::Positive,
-                Some(vec![sid_param, variant_param]),
-                None,
-            )
-        };
-        let diag_service = new_diag_service!(
-            db_builder,
-            vd_diag_comm,
-            vd_request,
-            vec![pos_response_variant_code],
-            vec![]
-        );
-
-        // Create base variant with pattern matching variant_code = 0
-        let base_variant = {
-            let matching_param_base =
-                db_builder.create_matching_parameter("0", diag_service, variant_param);
-            let variant_pattern_base =
-                db_builder.create_variant_pattern(&vec![matching_param_base]);
-            let base_diag_layer = db_builder.create_diag_layer(DiagLayerParams {
-                short_name: "BaseVariant",
-                com_param_refs: Some(vec![cp_ref]),
-                diag_services: Some(vec![vd_diag_service]),
-                state_charts: Some(vec![session_state_chart, security_state_chart]),
-                ..Default::default()
-            });
-            db_builder.create_variant(
-                base_diag_layer,
-                true,
-                Some(vec![variant_pattern_base]),
-                None,
-            )
-        };
-
-        // Create second variant with pattern matching variant_code = 1
-        let specific_variant = {
-            let matching_param_base =
-                db_builder.create_matching_parameter("1", diag_service, variant_param);
-            let variant_pattern_base =
-                db_builder.create_variant_pattern(&vec![matching_param_base]);
-            let base_diag_layer = db_builder.create_diag_layer(DiagLayerParams {
-                short_name: "SpecificVariant",
-                com_param_refs: Some(vec![cp_ref]),
-                diag_services: Some(vec![vd_diag_service]),
-                state_charts: Some(vec![session_state_chart, security_state_chart]),
-                ..Default::default()
-            });
-            db_builder.create_variant(
-                base_diag_layer,
-                false,
-                Some(vec![variant_pattern_base]),
-                None,
-            )
-        };
-
-        // we need multiple variants, hence cannot use the finish_db! macro,
-        // so we finish the db manually here.
-        let db = db_builder.finish(EcuDataParams {
-            ecu_name: "VariantDetectionEcu",
-            revision: "revision_1",
-            version: "1.0.0",
-            variants: Some(vec![base_variant, specific_variant]),
-            ..Default::default()
-        });
-
-        if fallback_to_base {
-            new_ecu_manager(db)
-        } else {
-            new_ecu_manager_no_base_fallback(db)
-        }
-    }
-
-    /// Creates an `EcuManager` with a service that uses a `PhysConst` parameter with a Normal DOP.
-    ///
-    /// Service layout (response):
-    ///   byte 0: SID (CODED-CONST)
-    ///   byte 1-2: DID (PHYS-CONST, Normal DOP, u16)
-    ///   byte 3: `data_param` (VALUE, u8)
-    #[allow(clippy::too_many_lines)]
-    fn create_ecu_manager_with_phys_const_normal_dop_service() -> (
-        super::EcuManager<DefaultSecurityPluginData>,
-        cda_interfaces::DiagComm,
-        u8,
-    ) {
-        let mut db_builder = EcuDataBuilder::new();
-        let u8_diag_type = db_builder.create_diag_coded_type_standard_length(8, DataType::UInt32);
-        let u16_diag_type = db_builder.create_diag_coded_type_standard_length(16, DataType::UInt32);
-        let compu_identical =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-
-        // Create Normal DOP for the PhysConst DID parameter
-        let did_dop = {
-            let did_dop_specific_data = db_builder
-                .create_normal_specific_dop_data(
-                    Some(compu_identical),
-                    Some(u16_diag_type),
-                    None,
-                    None,
-                    None,
-                    None,
-                )
-                .value_offset();
-            db_builder.create_dop(
-                *DopType::REGULAR,
-                Some("did_dop"),
-                None,
-                *SpecificDOPData::NormalDOP,
-                Some(did_dop_specific_data),
-            )
-        };
-
-        // Create Normal DOP for the VALUE data parameter
-        let data_dop = {
-            let data_dop_specific_data = db_builder
-                .create_normal_specific_dop_data(
-                    Some(compu_identical),
-                    Some(u8_diag_type),
-                    None,
-                    None,
-                    None,
-                    None,
-                )
-                .value_offset();
-            db_builder.create_dop(
-                *DopType::REGULAR,
-                Some("data_dop"),
-                None,
-                *SpecificDOPData::NormalDOP,
-                Some(data_dop_specific_data),
-            )
-        };
-
-        let sid = service_ids::READ_DATA_BY_IDENTIFIER + cda_interfaces::UDS_ID_RESPONSE_BITMASK;
-        let dc_name = "TestPhysConstNormalService";
-        let protocol_name = Protocol::default().to_string();
-        let protocol = db_builder.create_protocol(&protocol_name, None, None, None);
-        let diag_comm = new_diag_comm!(db_builder, dc_name, protocol);
-
-        // Request: SID (coded const) + DID (phys const)
-        let request = {
-            let sid_param =
-                create_sid_param!(db_builder, "sid", service_ids::READ_DATA_BY_IDENTIFIER);
-            let did_param = db_builder.create_phys_const_param("DID", Some("61840"), did_dop, 1, 0);
-            db_builder.create_request(Some(vec![sid_param, did_param]), None)
-        };
-
-        // Positive response: SID + DID (phys const) + data (value)
-        let pos_response = {
-            let sid_param = create_sid_param!(db_builder, "sid", sid);
-            let did_param = db_builder.create_phys_const_param("DID", Some("61840"), did_dop, 1, 0);
-            let data_param = db_builder.create_value_param("data_param", data_dop, 3, 0);
-            db_builder.create_response(
-                ResponseType::Positive,
-                Some(vec![sid_param, did_param, data_param]),
-                None,
-            )
-        };
-
-        let diag_service =
-            new_diag_service!(db_builder, diag_comm, request, vec![pos_response], vec![]);
-
-        let db = finish_db!(db_builder, protocol, vec![diag_service]);
-
-        let ecu_manager = new_ecu_manager(db);
-
-        let dc = cda_interfaces::DiagComm {
-            name: dc_name.to_owned(),
-            type_: DiagCommType::Data,
-            lookup_name: Some(dc_name.to_owned()),
-            subfunction_id: None,
-        };
-
-        (ecu_manager, dc, sid)
-    }
-
-    /// Creates an `EcuManager` with a service that uses a `PhysConst` parameter
-    /// with a Structure DOP.
-    ///
-    /// Service layout (request):
-    ///   byte 0: SID (CODED-CONST)
-    ///   byte 1-2: DID (PHYS-CONST, Normal DOP, u16)
-    ///   byte 3+: DREC (PHYS-CONST, Structure DOP with sub-params)
-    ///     sub-param1: u16 at byte 0
-    ///     sub-param2: u8 at byte 2
-    #[allow(clippy::too_many_lines)]
-    fn create_ecu_manager_with_phys_const_structure_dop_service() -> (
-        super::EcuManager<DefaultSecurityPluginData>,
-        cda_interfaces::DiagComm,
-        u8,
-    ) {
-        let mut db_builder = EcuDataBuilder::new();
-        let u8_diag_type = db_builder.create_diag_coded_type_standard_length(8, DataType::UInt32);
-        let u16_diag_type = db_builder.create_diag_coded_type_standard_length(16, DataType::UInt32);
-        let compu_identical =
-            db_builder.create_compu_method(datatypes::CompuCategory::Identical, None, None);
-
-        // Create Normal DOP for the PhysConst DID parameter
-        let did_dop = {
-            let did_dop_specific_data = db_builder
-                .create_normal_specific_dop_data(
-                    Some(compu_identical),
-                    Some(u16_diag_type),
-                    None,
-                    None,
-                    None,
-                    None,
-                )
-                .value_offset();
-            db_builder.create_dop(
-                *DopType::REGULAR,
-                Some("did_dop"),
-                None,
-                *SpecificDOPData::NormalDOP,
-                Some(did_dop_specific_data),
-            )
-        };
-
-        // Create Structure DOP for the PhysConst DREC parameter
-        let structure_dop = {
-            // Sub-param DOPs
-            let sub_param1_dop = {
-                let specific_data = db_builder
-                    .create_normal_specific_dop_data(
-                        Some(compu_identical),
-                        Some(u16_diag_type),
-                        None,
-                        None,
-                        None,
-                        None,
-                    )
-                    .value_offset();
-                db_builder.create_dop(
-                    *DopType::REGULAR,
-                    Some("sub_param1_dop"),
-                    None,
-                    *SpecificDOPData::NormalDOP,
-                    Some(specific_data),
-                )
-            };
-
-            let sub_param2_dop = {
-                let specific_data = db_builder
-                    .create_normal_specific_dop_data(
-                        Some(compu_identical),
-                        Some(u8_diag_type),
-                        None,
-                        None,
-                        None,
-                        None,
-                    )
-                    .value_offset();
-                db_builder.create_dop(
-                    *DopType::REGULAR,
-                    Some("sub_param2_dop"),
-                    None,
-                    *SpecificDOPData::NormalDOP,
-                    Some(specific_data),
-                )
-            };
-
-            // Create structure params
-            let sub_param1 = db_builder.create_value_param("sub_param1", sub_param1_dop, 0, 0);
-            let sub_param2 = db_builder.create_value_param("sub_param2", sub_param2_dop, 2, 0);
-
-            let structure = db_builder.create_structure(
-                Some(vec![sub_param1, sub_param2]),
-                Some(3), // byte_size: 2 bytes (u16) + 1 byte (u8) = 3 bytes
-                true,
-            );
-
-            db_builder.create_structure_dop("structure_dop", structure)
-        };
-
-        let protocol_name = Protocol::default().to_string();
-        let protocol = db_builder.create_protocol(&protocol_name, None, None, None);
-        let sid_request = service_ids::WRITE_DATA_BY_IDENTIFIER;
-        let sid_response =
-            service_ids::WRITE_DATA_BY_IDENTIFIER + cda_interfaces::UDS_ID_RESPONSE_BITMASK;
-        let dc_name = "TestPhysConstStructureService";
-        let diag_comm = new_diag_comm!(db_builder, dc_name, protocol);
-
-        // Request: SID (coded const) + DID (phys const, normal) + DREC (phys const, structure)
-        let request = {
-            let sid_param = create_sid_param!(db_builder, "sid", sid_request);
-            let did_param = db_builder.create_phys_const_param("DID", Some("61840"), did_dop, 1, 0);
-            let drec_param = db_builder.create_phys_const_param("DREC", None, structure_dop, 3, 0);
-            db_builder.create_request(Some(vec![sid_param, did_param, drec_param]), None)
-        };
-
-        // Positive response: SID + DID (phys const) + DREC (phys const, structure)
-        let pos_response = {
-            let sid_param = create_sid_param!(db_builder, "sid", sid_response);
-            let did_param = db_builder.create_phys_const_param("DID", Some("61840"), did_dop, 1, 0);
-            let drec_param = db_builder.create_phys_const_param("DREC", None, structure_dop, 3, 0);
-            db_builder.create_response(
-                ResponseType::Positive,
-                Some(vec![sid_param, did_param, drec_param]),
-                None,
-            )
-        };
-
-        let diag_service =
-            new_diag_service!(db_builder, diag_comm, request, vec![pos_response], vec![]);
-
-        let db = finish_db!(db_builder, protocol, vec![diag_service]);
-
-        let ecu_manager = new_ecu_manager(db);
-
-        let dc = cda_interfaces::DiagComm {
-            name: dc_name.to_owned(),
-            type_: DiagCommType::Configurations,
-            lookup_name: Some(dc_name.to_owned()),
-            subfunction_id: None,
-        };
-
-        (ecu_manager, dc, sid_response)
-    }
-
-    /// Whether the service's security `state_transition_ref` should use the
-    /// `Locked -> Extended` or `Extended -> Programming` transition.
-    #[derive(Copy, Clone)]
-    enum ServiceSecurityTransition {
-        /// Service references `LockedSecurity -> ExtendedSecurity`.
-        /// The default state (`LockedSecurity`) is also the transition source,
-        /// so it is always implicitly allowed.
-        LockedToExtended,
-        /// Service references `ExtendedSecurity -> ProgrammingSecurity`.
-        /// The default state (`LockedSecurity`) is NOT a transition source,
-        /// which means it can be rejected.
-        ExtendedToProgramming,
-    }
-
-    /// Helper function to create an ECU manager with services that have state transition refs.
-    ///
-    /// `security_transition` controls which security transition the service
-    /// references, which affects whether the default state ends up in the
-    /// allowed set.
-    fn create_ecu_manager_with_state_transitions(
-        security_transition: ServiceSecurityTransition,
-    ) -> (
-        super::EcuManager<DefaultSecurityPluginData>,
-        cda_interfaces::DiagComm,
-    ) {
-        let mut db_builder = EcuDataBuilder::new();
-        let protocol_name = Protocol::default().to_string();
-        let protocol = db_builder.create_protocol(&protocol_name, None, None, None);
-        let cp_ref = db_builder.create_com_param_ref(None, None, None, Some(protocol), None);
-
-        // Create security states
-        let locked_state = db_builder.create_state("LockedSecurity", None);
-        let extended_state = db_builder.create_state("ExtendedSecurity", None);
-        let programming_state = db_builder.create_state("ProgrammingSecurity", None);
-
-        // Create session states
-        let default_session_state = db_builder.create_state("DefaultSession", None);
-        let extended_session_state = db_builder.create_state("ExtendedSession", None);
-        let programming_session_state = db_builder.create_state("ProgrammingSession", None);
-
-        // Create state transitions for session
-        let default_to_extended_session = db_builder.create_state_transition(
-            "DefaultToExtended",
-            Some("DefaultSession"),
-            Some("ExtendedSession"),
-        );
-        let extended_to_programming_session = db_builder.create_state_transition(
-            "ExtendedToProgramming",
-            Some("ExtendedSession"),
-            Some("ProgrammingSession"),
-        );
-
-        // Create state transitions for security
-        let locked_to_extended_transition = db_builder.create_state_transition(
-            "LockedToExtended",
-            Some("LockedSecurity"),
-            Some("ExtendedSecurity"),
-        );
-
-        let extended_to_programming_transition = db_builder.create_state_transition(
-            "ExtendedToProgramming",
-            Some("ExtendedSecurity"),
-            Some("ProgrammingSecurity"),
-        );
-
-        // Create state chart for security
-        let security_state_chart = db_builder.create_state_chart(
-            "SecurityAccess",
-            Some(semantics::SECURITY),
-            Some(vec![
-                locked_to_extended_transition,
-                extended_to_programming_transition,
-            ]),
-            Some("LockedSecurity"),
-            Some(vec![locked_state, extended_state, programming_state]),
-        );
-
-        // Create state chart for session
-        let session_state_chart = db_builder.create_state_chart(
-            "Session",
-            Some(semantics::SESSION),
-            Some(vec![
-                default_to_extended_session,
-                extended_to_programming_session,
-            ]),
-            Some("DefaultSession"),
-            Some(vec![
-                default_session_state,
-                extended_session_state,
-                programming_session_state,
-            ]),
-        );
-
-        // Choose which security transition the service references
-        let service_security_transition = match security_transition {
-            ServiceSecurityTransition::LockedToExtended => locked_to_extended_transition,
-            ServiceSecurityTransition::ExtendedToProgramming => extended_to_programming_transition,
-        };
-
-        // Create state transition refs for the service
-        let state_transition_ref =
-            db_builder.create_state_transition_ref(service_security_transition);
-        let session_transition_ref =
-            db_builder.create_state_transition_ref(default_to_extended_session);
-
-        // Create precondition state ref - service requires Programming
-        let precondition_ref = db_builder.create_pre_condition_state_ref(programming_state);
-
-        // Create a service with state transition refs
-        let sid = service_ids::WRITE_DATA_BY_IDENTIFIER;
-        let dc_name = "TestServiceWithStateTransitions";
-
-        let diag_comm = db_builder.create_diag_comm(DiagCommParams {
-            short_name: dc_name,
-            pre_condition_state_refs: Some(vec![precondition_ref]),
-            state_transition_refs: Some(vec![session_transition_ref, state_transition_ref]),
-            protocols: Some(vec![protocol]),
-            ..Default::default()
-        });
-
-        let request = create_sid_only_request!(db_builder, sid);
-        let diag_service = new_diag_service!(db_builder, diag_comm, request, vec![], vec![]);
-
-        let diag_layer = db_builder.create_diag_layer(DiagLayerParams {
-            short_name: "TestVariantDiagLayer",
-            com_param_refs: Some(vec![cp_ref]),
-            diag_services: Some(vec![diag_service]),
-            state_charts: Some(vec![session_state_chart, security_state_chart]),
-            ..Default::default()
-        });
-
-        let variant = db_builder.create_variant(diag_layer, true, None, None);
-        let db = db_builder.finish(EcuDataParams {
-            revision: "revision_1",
-            version: "1.0.0",
-            variants: Some(vec![variant]),
-            ..Default::default()
-        });
-
-        let dc = cda_interfaces::DiagComm {
-            name: dc_name.to_owned(),
-            type_: DiagCommType::Configurations,
-            lookup_name: Some(dc_name.to_owned()),
-            subfunction_id: None,
-        };
-        (new_ecu_manager(db), dc)
-    }
-
-    /// Helper that builds an ECU manager whose variant has a service with a
-    /// `ProgrammingSecurity` precondition **and** a functional group containing
-    /// the same service. Returns `(ecu_manager, diag_comm, sid)`.
-    fn create_ecu_manager_with_preconditions_and_functional_group() -> (
-        super::EcuManager<DefaultSecurityPluginData>,
-        cda_interfaces::DiagComm,
-        u8,
-    ) {
-        let mut db_builder = EcuDataBuilder::new();
-        let protocol_name = Protocol::default().to_string();
-        let protocol = db_builder.create_protocol(&protocol_name, None, None, None);
-        let cp_ref = db_builder.create_com_param_ref(None, None, None, Some(protocol), None);
-
-        let locked_state = db_builder.create_state("LockedSecurity", None);
-        let extended_state = db_builder.create_state("ExtendedSecurity", None);
-        let programming_state = db_builder.create_state("ProgrammingSecurity", None);
-
-        let default_session_state = db_builder.create_state("DefaultSession", None);
-        let extended_session_state = db_builder.create_state("ExtendedSession", None);
-        let programming_session_state = db_builder.create_state("ProgrammingSession", None);
-
-        let locked_to_extended = db_builder.create_state_transition(
-            "LockedToExtended",
-            Some("LockedSecurity"),
-            Some("ExtendedSecurity"),
-        );
-        let extended_to_programming = db_builder.create_state_transition(
-            "ExtendedToProgramming",
-            Some("ExtendedSecurity"),
-            Some("ProgrammingSecurity"),
-        );
-        let default_to_extended_session = db_builder.create_state_transition(
-            "DefaultToExtended",
-            Some("DefaultSession"),
-            Some("ExtendedSession"),
-        );
-        let extended_to_programming_session = db_builder.create_state_transition(
-            "ExtendedToProgramming",
-            Some("ExtendedSession"),
-            Some("ProgrammingSession"),
-        );
-
-        let security_state_chart = db_builder.create_state_chart(
-            "SecurityAccess",
-            Some(semantics::SECURITY),
-            Some(vec![locked_to_extended, extended_to_programming]),
-            Some("LockedSecurity"),
-            Some(vec![locked_state, extended_state, programming_state]),
-        );
-        let session_state_chart = db_builder.create_state_chart(
-            "Session",
-            Some(semantics::SESSION),
-            Some(vec![
-                default_to_extended_session,
-                extended_to_programming_session,
-            ]),
-            Some("DefaultSession"),
-            Some(vec![
-                default_session_state,
-                extended_session_state,
-                programming_session_state,
-            ]),
-        );
-
-        let precondition_ref = db_builder.create_pre_condition_state_ref(programming_state);
-        let sid = service_ids::WRITE_DATA_BY_IDENTIFIER;
-        let dc_name = "TestFGService";
-
-        // Service for the functional group diag layer
-        let fg_diag_comm = db_builder.create_diag_comm(DiagCommParams {
-            short_name: dc_name,
-            pre_condition_state_refs: Some(vec![precondition_ref]),
-            protocols: Some(vec![protocol]),
-            ..Default::default()
-        });
-        let fg_request = create_sid_only_request!(db_builder, sid);
-        let fg_service = new_diag_service!(db_builder, fg_diag_comm, fg_request, vec![], vec![]);
-        let fg_layer = db_builder.create_diag_layer(DiagLayerParams {
-            short_name: "TestFunctionalGroup",
-            diag_services: Some(vec![fg_service]),
-            ..Default::default()
-        });
-        let functional_group = db_builder.create_functional_group(fg_layer, None);
-
-        // Same service for the variant diag layer
-        let variant_diag_comm = db_builder.create_diag_comm(DiagCommParams {
-            short_name: dc_name,
-            pre_condition_state_refs: Some(vec![precondition_ref]),
-            protocols: Some(vec![protocol]),
-            ..Default::default()
-        });
-        let variant_request = create_sid_only_request!(db_builder, sid);
-        let variant_service = new_diag_service!(
-            db_builder,
-            variant_diag_comm,
-            variant_request,
-            vec![],
-            vec![]
-        );
-        let variant_layer = db_builder.create_diag_layer(DiagLayerParams {
-            short_name: TEST_DIAG_LAYER,
-            com_param_refs: Some(vec![cp_ref]),
-            diag_services: Some(vec![variant_service]),
-            state_charts: Some(vec![session_state_chart, security_state_chart]),
-            ..Default::default()
-        });
-        let variant = db_builder.create_variant(variant_layer, true, None, None);
-
-        let db = db_builder.finish(EcuDataParams {
-            ecu_name: "TestEcu",
-            revision: "1",
-            version: "1.0.0",
-            variants: Some(vec![variant]),
-            functional_groups: Some(vec![functional_group]),
-            ..Default::default()
-        });
-
-        let dc = cda_interfaces::DiagComm {
-            name: dc_name.to_owned(),
-            type_: DiagCommType::Configurations,
-            lookup_name: Some(dc_name.to_owned()),
-            subfunction_id: None,
-        };
-        (new_ecu_manager(db), dc, sid)
-    }
-
-    fn create_ecu_manager_with_length_key_request_service()
-    -> (super::EcuManager<DefaultSecurityPluginData>, DiagComm, u8) {
-        let mut db_builder = EcuDataBuilder::new();
-        let protocol_name = Protocol::default().to_string();
-        let protocol = db_builder.create_protocol(&protocol_name, None, None, None);
-        let compu_identical = db_builder.create_compu_method(CompuCategory::Identical, None, None);
-
-        let u8_diag_type = db_builder.create_diag_coded_type_standard_length(8, DataType::UInt32);
-        let u16_diag_type = db_builder.create_diag_coded_type_standard_length(16, DataType::UInt32);
-
-        let length_key_dop =
-            db_builder.create_regular_normal_dop("lk_dop", u8_diag_type, compu_identical);
-        let value_dop =
-            db_builder.create_regular_normal_dop("val_dop", u16_diag_type, compu_identical);
-
-        let sid = service_ids::WRITE_DATA_BY_IDENTIFIER;
-        let dc_name = "TestLengthKeyReqService";
-        let diag_comm = new_diag_comm!(db_builder, dc_name, protocol);
-
-        let request = {
-            let sid_param = create_sid_param!(db_builder, sid);
-            let lk_param =
-                db_builder.create_length_key_param("length_indicator", length_key_dop, 1, 0);
-            let val_param = db_builder.create_value_param("value_param", value_dop, 2, 0);
-            db_builder.create_request(Some(vec![sid_param, lk_param, val_param]), None)
-        };
-
-        let pos_response = {
-            let sid_param = create_sid_param!(
-                db_builder,
-                "pos_sid",
-                sid.saturating_add(UDS_ID_RESPONSE_BITMASK)
-            );
-            db_builder.create_response(ResponseType::Positive, Some(vec![sid_param]), None)
-        };
-
-        let diag_service =
-            new_diag_service!(db_builder, diag_comm, request, vec![pos_response], vec![]);
-        let db = finish_db!(db_builder, protocol, vec![diag_service]);
-        (
-            new_ecu_manager(db),
-            DiagComm::new(dc_name, DiagCommType::Configurations),
-            sid,
-        )
-    }
-
-    fn create_ecu_manager_with_param_length_info_service()
-    -> (super::EcuManager<DefaultSecurityPluginData>, DiagComm, u8) {
-        const LEN_KEY: &str = "len_key";
-        const VAR_DATA: &str = "var_data";
-
-        let mut db_builder = EcuDataBuilder::new();
-        let protocol_name = Protocol::default().to_string();
-        let protocol = db_builder.create_protocol(&protocol_name, None, None, None);
-        let compu_identical = db_builder.create_compu_method(CompuCategory::Identical, None, None);
-
-        let len_key_diag_type =
-            db_builder.create_diag_coded_type_standard_length(8, DataType::UInt32);
-        let var_data_diag_type =
-            db_builder.create_diag_coded_type_param_length_info(LEN_KEY, DataType::ByteField);
-
-        let len_key_dop =
-            db_builder.create_regular_normal_dop("len_key_dop", len_key_diag_type, compu_identical);
-        let var_data_dop = db_builder.create_regular_normal_dop(
-            "var_data_dop",
-            var_data_diag_type,
-            compu_identical,
-        );
-
-        let sid = service_ids::WRITE_DATA_BY_IDENTIFIER;
-        let dc_name = "TestParamLengthInfoService";
-        let diag_comm = new_diag_comm!(db_builder, dc_name, protocol);
-
-        let request = {
-            let sid_param = create_sid_param!(db_builder, sid);
-            let len_key_param = db_builder.create_length_key_param(LEN_KEY, len_key_dop, 1, 0);
-            let var_data_param = db_builder.create_value_param(VAR_DATA, var_data_dop, 2, 0);
-            db_builder.create_request(Some(vec![sid_param, len_key_param, var_data_param]), None)
-        };
-
-        let pos_response = {
-            let pos_sid_param = create_sid_param!(
-                db_builder,
-                "pos_sid",
-                sid.saturating_add(UDS_ID_RESPONSE_BITMASK)
-            );
-            let len_key_param = db_builder.create_length_key_param(LEN_KEY, len_key_dop, 1, 0);
-            let var_data_param = db_builder.create_value_param(VAR_DATA, var_data_dop, 2, 0);
-            db_builder.create_response(
-                ResponseType::Positive,
-                Some(vec![pos_sid_param, len_key_param, var_data_param]),
-                None,
-            )
-        };
-
-        let diag_service =
-            new_diag_service!(db_builder, diag_comm, request, vec![pos_response], vec![]);
-        let db = finish_db!(db_builder, protocol, vec![diag_service]);
-        (
-            new_ecu_manager(db),
-            DiagComm::new(dc_name, DiagCommType::Configurations),
-            sid,
-        )
-    }
-
-    // Models the pattern from ISO 22901-1 §7.4.8 (readMemoryByAddress):
-    // one parameter determines the length of the next, and the parameter
-    // that comes *after* the variable-length data has no BYTE-POSITION in
-    // the ODX because its position is unknown until runtime.
-    //
-    // Layout (request & positive response):
-    //   byte 0     : SID       (coded const, 8 bit)
-    //   byte 1     : len_key   (LENGTH-KEY param, u8)
-    //   byte 2     : var_data  (PARAM-LENGTH-INFO, `len_key` bytes)
-    //   byte 2 + N : suffix    (value param, u16 - BYTE-POSITION omitted)
-    fn create_ecu_manager_with_trailing_param_after_param_length_info_service()
-    -> (super::EcuManager<DefaultSecurityPluginData>, DiagComm, u8) {
-        const LEN_KEY: &str = "len_key";
-        const VAR_DATA: &str = "var_data";
-
-        let mut db_builder = EcuDataBuilder::new();
-        let protocol_name = Protocol::default().to_string();
-        let protocol = db_builder.create_protocol(&protocol_name, None, None, None);
-        let compu_identical = db_builder.create_compu_method(CompuCategory::Identical, None, None);
-
-        // diag coded types
-        let u8_dct = db_builder.create_diag_coded_type_standard_length(8, DataType::UInt32);
-        let u16_dct = db_builder.create_diag_coded_type_standard_length(16, DataType::UInt32);
-        let var_dct =
-            db_builder.create_diag_coded_type_param_length_info(LEN_KEY, DataType::ByteField);
-
-        // DOPs
-        let len_key_dop =
-            db_builder.create_regular_normal_dop("len_key_dop", u8_dct, compu_identical);
-        let var_data_dop =
-            db_builder.create_regular_normal_dop("var_data_dop", var_dct, compu_identical);
-        let suffix_dop =
-            db_builder.create_regular_normal_dop("suffix_dop", u16_dct, compu_identical);
-
-        let sid = service_ids::WRITE_DATA_BY_IDENTIFIER;
-        let dc_name = "TestTrailingParamAfterPLI";
-        let diag_comm = new_diag_comm!(db_builder, dc_name, protocol);
-
-        let request = {
-            let sid_param = create_sid_param!(db_builder, sid);
-            let lk_param = db_builder.create_length_key_param(LEN_KEY, len_key_dop, 1, 0);
-            let var_param = db_builder.create_value_param(VAR_DATA, var_data_dop, 2, 0);
-            // Per spec: BYTE-POSITION omitted - position depends on runtime length of var_data
-            let suffix_param = db_builder.create_value_param_no_byte_pos("suffix", suffix_dop);
-            db_builder.create_request(
-                Some(vec![sid_param, lk_param, var_param, suffix_param]),
-                None,
-            )
-        };
-        let pos_response = {
-            let pos_sid_param = create_sid_param!(
-                db_builder,
-                "pos_sid",
-                sid.saturating_add(UDS_ID_RESPONSE_BITMASK)
-            );
-            let lk_param = db_builder.create_length_key_param(LEN_KEY, len_key_dop, 1, 0);
-            let var_param = db_builder.create_value_param(VAR_DATA, var_data_dop, 2, 0);
-            let suffix_param = db_builder.create_value_param_no_byte_pos("suffix", suffix_dop);
-            db_builder.create_response(
-                ResponseType::Positive,
-                Some(vec![pos_sid_param, lk_param, var_param, suffix_param]),
-                None,
-            )
-        };
-
-        let diag_service =
-            new_diag_service!(db_builder, diag_comm, request, vec![pos_response], vec![]);
-        let db = finish_db!(db_builder, protocol, vec![diag_service]);
-        (
-            new_ecu_manager(db),
-            DiagComm::new(dc_name, DiagCommType::Configurations),
-            sid,
-        )
-    }
-
-    /// Creates an ECU manager whose database contains a routine control service with the
-    /// following request structure:
-    /// - SID: 0x31 (Routine Control)
-    /// - Sub-function: 0x03 (8-bit, at byte position 1)
-    /// - Routine ID: 0x0A5C (16-bit, at byte positions 2-3)
-    fn create_ecu_manager_with_routine_control_service()
-    -> super::EcuManager<DefaultSecurityPluginData> {
-        const SERVICE_ID: u8 = 0x31;
-        const ROUTINE_ID: u16 = 0x0A5C;
-        const SERVICE_NAME: &str = "Test";
-
-        let mut db_builder = EcuDataBuilder::new();
-        let protocol_name = Protocol::default().to_string();
-        let protocol = db_builder.create_protocol(&protocol_name, None, None, None);
-
-        // Create the SID parameter
-        let sid_param = db_builder.create_coded_const_param(
-            "SID_RQ",
-            &SERVICE_ID.to_string(),
-            0,
-            0,
-            8,
-            DataType::UInt32,
-        );
-
-        // Create the subfunction parameter
-        let subfunction_param = db_builder.create_coded_const_param(
-            "RoutineControlType",
-            &subfunction_ids::routine::REQUEST_RESULTS.to_string(),
-            1,
-            0,
-            8,
-            DataType::UInt32,
-        );
-
-        // Create the routine ID parameter
-        let routine_id_param = db_builder.create_coded_const_param(
-            "RoutineIdentifier",
-            &ROUTINE_ID.to_string(),
-            2,
-            0,
-            16,
-            DataType::UInt32,
-        );
-
-        // Create the request with all three parameters
-        let request = db_builder.create_request(
-            Some(vec![sid_param, subfunction_param, routine_id_param]),
-            None,
-        );
-
-        // Create the DiagComm
-        let diag_comm = db_builder.create_diag_comm(DiagCommParams {
-            short_name: SERVICE_NAME,
-            diag_class_type: DiagClassType::START_COMM,
-            protocols: Some(vec![protocol]),
-            ..Default::default()
-        });
-
-        // Create the DiagService
-        let diag_service = new_diag_service!(db_builder, diag_comm, request, vec![], vec![]);
-        let db = finish_db!(db_builder, protocol, vec![diag_service]);
-        new_ecu_manager(db)
     }
 
     #[tokio::test]
@@ -10038,188 +7149,6 @@ mod tests {
             "expected InvalidDatabase error"
         );
     }
-
-    /// Builds a deep, mixed-type hierarchy and asserts that every level is
-    /// resolved.
-    ///
-    /// ```text
-    /// Variant("RootVariant")
-    /// +-- Variant("InnerVariant")
-    /// |   \-- FunctionalGroup("FgLayer")
-    /// |       \-- EcuSharedData("SharedInFg")
-    /// +-- Protocol("Proto")
-    /// |   \-- Protocol("ParentProto")
-    /// \-- EcuSharedData("TopShared")
-    /// ```
-    ///
-    /// Expected collected layers (order is stack-based, not guaranteed):
-    ///   `TopShared`, Proto, `ParentProto`, `InnerVariant`, `FgLayer`, `SharedInFg`
-    #[test]
-    fn test_parent_ref_recursive_mixed_hierarchy() {
-        let mut b = EcuDataBuilder::new();
-
-        // - leaf: EcuSharedData inside a FunctionalGroup -
-        let shared_in_fg_dl = b.create_diag_layer(DiagLayerParams {
-            short_name: "SharedInFg",
-            ..Default::default()
-        });
-        let esd_in_fg = b.create_ecu_shared_data(shared_in_fg_dl);
-        let esd_in_fg_pr = b.create_parent_ref(
-            DataFormatParentRefType::EcuSharedData,
-            Some(DataFormatParentRefType::tag_as_ecu_shared_data(esd_in_fg)),
-        );
-
-        let fg_dl = b.create_diag_layer(DiagLayerParams {
-            short_name: "FgLayer",
-            ..Default::default()
-        });
-        let fg = b.create_functional_group(fg_dl, Some(vec![esd_in_fg_pr]));
-        let fg_pr = b.create_parent_ref(
-            DataFormatParentRefType::FunctionalGroup,
-            Some(DataFormatParentRefType::tag_as_functional_group(fg)),
-        );
-
-        // - inner Variant whose parent-ref is the FunctionalGroup -
-        let inner_variant_dl = b.create_diag_layer(DiagLayerParams {
-            short_name: "InnerVariant",
-            ..Default::default()
-        });
-        let inner_variant = b.create_variant(inner_variant_dl, false, None, Some(vec![fg_pr]));
-        let variant_pr = b.create_parent_ref(
-            DataFormatParentRefType::Variant,
-            Some(DataFormatParentRefType::tag_as_variant(inner_variant)),
-        );
-
-        // - Protocol with a parent protocol -
-        let parent_proto = b.create_protocol("ParentProto", None, None, None);
-        let parent_proto_pr = b.create_parent_ref(
-            DataFormatParentRefType::Protocol,
-            Some(DataFormatParentRefType::tag_as_protocol(parent_proto)),
-        );
-        let proto = b.create_protocol("Proto", None, None, Some(vec![parent_proto_pr]));
-        let proto_pr = b.create_parent_ref(
-            DataFormatParentRefType::Protocol,
-            Some(DataFormatParentRefType::tag_as_protocol(proto)),
-        );
-
-        // - top-level EcuSharedData sibling -
-        let top_shared_dl = b.create_diag_layer(DiagLayerParams {
-            short_name: "TopShared",
-            ..Default::default()
-        });
-        let top_esd = b.create_ecu_shared_data(top_shared_dl);
-        let top_esd_pr = b.create_parent_ref(
-            DataFormatParentRefType::EcuSharedData,
-            Some(DataFormatParentRefType::tag_as_ecu_shared_data(top_esd)),
-        );
-
-        // - root variant carrying all three sibling parent-refs -
-        let root_dl = b.create_diag_layer(DiagLayerParams {
-            short_name: "RootVariant",
-            ..Default::default()
-        });
-        let root = b.create_variant(
-            root_dl,
-            true,
-            None,
-            Some(vec![variant_pr, proto_pr, top_esd_pr]),
-        );
-        let db = b.finish(EcuDataParams {
-            ecu_name: "TestEcu",
-            revision: "1",
-            version: "1.0.0",
-            variants: Some(vec![root]),
-            ..Default::default()
-        });
-
-        let ecu_data = db.ecu_data().unwrap();
-        let variant = ecu_data.variants().unwrap().get(0);
-        let parent_refs = variant.parent_refs().unwrap();
-
-        let names: Vec<_> = super::EcuManager::<DefaultSecurityPluginData>
-            ::get_parent_ref_diag_layers_with_refs_recursive(
-            parent_refs.iter().map(datatypes::ParentRef),
-        )
-            .into_iter()
-            .filter_map(|(_, dl)| dl.short_name().map(str::to_owned))
-            .collect();
-
-        // every layer from every level must be present
-        for expected in [
-            "TopShared",
-            "Proto",
-            "ParentProto",
-            "InnerVariant",
-            "FgLayer",
-            "SharedInFg",
-        ] {
-            assert!(
-                names.contains(&expected.to_owned()),
-                "Missing expected layer {expected:?}, got {names:?}"
-            );
-        }
-        assert_eq!(names.len(), 6, "Unexpected extra layers: {names:?}");
-    }
-
-    /// Test `lookup_service_by_request_prefix` with a routine control service.
-    #[test]
-    fn test_lookup_service_by_request_prefix_routine_control() {
-        const SERVICE_ID: u8 = 0x31;
-        const SERVICE_NAME: &str = "Test";
-
-        fn assert_success(result: Result<Vec<DiagComm>, DiagServiceError>) {
-            assert!(result.is_ok(), "Expected successful lookup");
-            let services = result.unwrap();
-            assert_eq!(services.len(), 1, "Expected exactly one matching service");
-            assert_eq!(
-                services
-                    .first()
-                    .expect("Expected at least one service")
-                    .lookup_name
-                    .as_ref()
-                    .expect("Expected lookup name in DiagComm to be set"),
-                SERVICE_NAME,
-                "Expected service name to match"
-            );
-        }
-
-        let ecu_manager = create_ecu_manager_with_routine_control_service();
-
-        // Lookup with complete prefix (all 4 bytes)
-        let full_prefix = vec![
-            SERVICE_ID,
-            subfunction_ids::routine::REQUEST_RESULTS,
-            0x0A,
-            0x5C,
-        ];
-        let result = ecu_manager.lookup_diagcomms_by_request_prefix(&full_prefix);
-        assert_success(result);
-
-        // Lookup with partial request
-        // (first 3 bytes - SID + subfunction + first byte of routine ID)
-        let partial_prefix = vec![SERVICE_ID, subfunction_ids::routine::REQUEST_RESULTS, 0x0A];
-        let result = ecu_manager.lookup_diagcomms_by_request_prefix(&partial_prefix);
-        assert_success(result);
-
-        // Lookup with wrong subfunction
-        let wrong_subfunction = vec![SERVICE_ID, subfunction_ids::routine::STOP, 0x0A, 0x5C];
-        let result = ecu_manager.lookup_diagcomms_by_request_prefix(&wrong_subfunction);
-        assert!(
-            result.is_err(),
-            "Expected lookup to fail with wrong subfunction"
-        );
-
-        // Lookup with empty prefix
-        let result = ecu_manager.lookup_diagcomms_by_request_prefix(&[]);
-        assert!(result.is_err(), "Expected lookup to fail with empty prefix");
-        match result.unwrap_err() {
-            DiagServiceError::NotFound { .. } => {
-                // Expected error type
-            }
-            other => panic!("Expected NotFound error, got: {other:?}"),
-        }
-    }
-
     /// Build an `EcuManager` whose database contains `RoutineControl` services for a
     /// routine named `routine_name`.  `subfunctions` controls which subfunction bytes
     /// (0x01 = Start, 0x02 = Stop, 0x03 = `RequestResults`) are included.
